@@ -13,21 +13,19 @@
 # limitations under the License.
 
 """An ONNX to XLA compiler by JAX-tracing a Numpy-backed ONNX interpreter."""
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
 
-from cStringIO import StringIO
+from io import BytesIO
 from functools import partial
 import hashlib
+import urllib.request
 import sys
 
+import numpy as np
 import onnx
 from onnx import numpy_helper
-from onnx import onnx_pb2
-import urllib
 
-import jax.numpy as np
+import jax.numpy as jnp
+import jax.ops
 from jax import jit, grad
 from jax import lax
 
@@ -36,7 +34,7 @@ def _asarray(proto):
   return numpy_helper.to_array(proto).reshape(tuple(proto.dims))
 
 
-attr_types = dict(onnx_pb2.AttributeProto.AttributeType.items())
+attr_types = dict(onnx.AttributeProto.AttributeType.items())
 attribute_handlers = {
     attr_types['FLOAT']: lambda a: a.f,
     attr_types['INT']: lambda a: a.i,
@@ -55,7 +53,7 @@ def onnx_maxpool(x, kernel_shape, pads=None, strides=None):
   dims = prefix + tuple(kernel_shape)
   pads = tuple(pads) if pads else [0] * len(kernel_shape)
   strides = (prefix + tuple(strides)) if strides else [1] * len(kernel_shape)
-  return [lax.reduce_window(x, -np.inf, lax.max, dims, strides, 'VALID')]
+  return [lax.reduce_window(x, -jnp.inf, lax.max, dims, strides, 'VALID')]
 
 
 def onnx_conv(x, w, b=0, group=1, kernel_shape=None, pads=None, strides=None,
@@ -65,7 +63,7 @@ def onnx_conv(x, w, b=0, group=1, kernel_shape=None, pads=None, strides=None,
   kernel_shape = kernel_shape or w.shape
   strides = strides or [1] * (w.ndim - 2)
   if auto_pad:
-    auto_pad = 'SAME' if auto_pad.startswith('SAME') else 'VALID'
+    auto_pad = 'SAME' if auto_pad.startswith(b'SAME') else 'VALID'
     pads = lax.padtype_to_pads(x.shape[2:], w.shape[2:], strides, auto_pad)
   else:
     pads = pads or [0] * (w.ndim - 2)
@@ -80,9 +78,9 @@ def onnx_add(a, b, axis=None, broadcast=True):
   if broadcast:
     axis = (a.dim - b.ndim) if axis is None else axis % a.ndim
     assert a.shape[axis:][:b.ndim] == b.shape
-    b_shape = np.ones(a.ndim, dtype='int64').copy()
+    b_shape = np.ones(a.ndim, dtype='int64')
     b_shape[axis:axis + b.ndim] = b.shape
-    b = np.reshape(b, b_shape)
+    b = jnp.reshape(b, b_shape)
   return [a + b]
 
 
@@ -90,10 +88,10 @@ onnx_ops = {
     'Add': onnx_add,
     'Constant': lambda value: [value],
     'Conv': onnx_conv,
-    'MatMul': lambda x, y: [np.matmul(x, y)],
+    'MatMul': lambda x, y: [jnp.matmul(x, y)],
     'MaxPool': onnx_maxpool,
-    'Relu': lambda x: [np.maximum(x, 0)],
-    'Reshape': lambda x, shape: [np.reshape(x, shape)],
+    'Relu': lambda x: [jnp.maximum(x, 0)],
+    'Reshape': lambda x, shape: [jnp.reshape(x, shape)],
 }
 
 
@@ -119,21 +117,21 @@ if __name__ == "__main__":
   if hashlib.md5(download).hexdigest() != 'bc8ad9bd19c5a058055dc18d0f089dad':
     print("onnx file checksum mismatch")
     sys.exit(1)
-  model = onnx.load(StringIO(download))
+  model = onnx.load(BytesIO(download))
 
   predict = lambda inputs: interpret_onnx(model.graph, inputs)[0]
 
   # Run inference in Numpy-backed interpreter
   print("interpreted:")
-  print(predict(np.ones((1, 1, 28, 28))))
+  print(predict(jnp.ones((1, 1, 28, 28))))
 
   # JIT compile to XLA device, run inference on device
   compiled_predict = jit(predict)
   print("compiled:")
-  print(compiled_predict(np.ones((1, 1, 28, 28))))
+  print(compiled_predict(jnp.ones((1, 1, 28, 28))))
 
   # The interpreter is differentiable too! Even the compiled one:
-  fun = lambda inputs: np.sum(compiled_predict(inputs))
+  fun = lambda inputs: jnp.sum(compiled_predict(inputs))
   print("a derivative with respect to inputs:")
-  print(grad(fun)(np.ones((1, 1, 28, 28)))[..., :3, :3])
+  print(grad(fun)(jnp.ones((1, 1, 28, 28)))[..., :3, :3])
 
