@@ -12,14 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
 
 from .tree_util import (build_tree, tree_flatten, tree_unflatten,
                         treedef_is_leaf)
 from . import linear_util as lu
-from .util import safe_map, unzip2, partial, curry
+from .util import safe_map, unzip2, partial, curry, WrapHashably, Hashable
+from .core import unit
 
 map = safe_map
 
@@ -73,3 +71,29 @@ def flatten_fun_nokwargs2(in_tree, *args_flat):
   ans_flat, ans_tree = tree_flatten(ans)
   aux_flat, aux_tree = tree_flatten(aux)
   yield (ans_flat, aux_flat), (ans_tree, aux_tree)
+
+def argnums_partial(f, dyn_argnums, args):
+  if isinstance(dyn_argnums, int):
+    dyn_argnums = (dyn_argnums,)
+  else:
+    dyn_argnums = tuple(dyn_argnums)
+  fixed_args = tuple([unit if i in dyn_argnums else wrap_hashably(arg)
+                      for i, arg in enumerate(args)])
+  dyn_args = tuple(args[i] for i in dyn_argnums)
+  return _argnums_partial(f, dyn_argnums, fixed_args), dyn_args
+
+def wrap_hashably(arg):
+  try:
+    hash(arg)
+  except TypeError:
+    return WrapHashably(arg)  # e.g. ndarrays, DeviceArrays
+  else:
+    return Hashable(arg)
+
+@lu.transformation
+def _argnums_partial(dyn_argnums, fixed_args, *dyn_args, **kwargs):
+  args = [None if arg is unit else arg.val for arg in fixed_args]
+  for i, arg in zip(dyn_argnums, dyn_args):
+    args[i] = arg
+  ans = yield args, kwargs
+  yield ans
