@@ -12,24 +12,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
 
 from functools import partial
-import warnings
-
-import numpy as onp
+from typing import Callable, Dict
 
 from .. import core
 from .. import linear_util as lu
 from ..core import Trace, Tracer, Primitive, new_master
-from ..abstract_arrays import ShapedArray, ConcreteArray, raise_to_shaped
-from ..util import safe_map, safe_zip, unzip2, unzip3, partialmethod, prod
-from ..lib import xla_bridge as xb
+from ..abstract_arrays import ShapedArray, raise_to_shaped
+from ..util import safe_map, safe_zip, unzip2, unzip3
 from . import partial_eval as pe
-from . import batching
-from . import pxla
 
 map = safe_map
 zip = safe_zip
@@ -69,7 +61,7 @@ not_sharded = None
 
 class PapplyTracer(Tracer):
   def __init__(self, trace, name, axis_size, val, axis):
-    self.trace = trace
+    self._trace = trace
     self.name = name
     self.axis_size = axis_size
     self.val = val
@@ -118,7 +110,7 @@ class PapplyTrace(Trace):
       val_out, axis_out = rule(name, size, vals, axes, **params)
       return PapplyTracer(self, name, size, val_out, axis_out)
 
-  def process_call(self, call_primitive, f, tracers, params):
+  def process_call(self, call_primitive, f: lu.WrappedFun, tracers, params):
     if call_primitive in pe.map_primitives:
       return self.process_map(call_primitive, f, tracers, params)
     names, vals, axes = unzip3((t.name, t.val, t.axis) for t in tracers)
@@ -141,8 +133,8 @@ class PapplyTrace(Trace):
       return PapplyTracer(trace, name, size, x, axis)
     return val, todo
 
-  def process_map(self, map_primitive, f, tracers, params):
+  def process_map(self, map_primitive, f :lu.WrappedFun, tracers, params):
     raise NotImplementedError  # TODO(mattjj,frostig)
 
 
-papply_primitive_rules = {}
+papply_primitive_rules: Dict[core.Primitive, Callable] = {}
