@@ -12,14 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
 
 import itertools as it
-from collections import namedtuple, Counter, defaultdict
+from collections import namedtuple
 import contextlib
 import threading
+from typing import Callable, Dict, Sequence, Set
 from weakref import ref
 
 import numpy as onp
@@ -27,7 +25,8 @@ import numpy as onp
 from .. import core
 from .. import linear_util as lu
 from ..abstract_arrays import ShapedArray, ConcreteArray, raise_to_shaped
-from ..util import unzip2, safe_zip, safe_map, toposort, partial, split_list
+from ..util import (unzip2, safe_zip, safe_map, toposort, partial, split_list,
+                    wrap_name, cache)
 from ..core import (Trace, Tracer, new_master, Jaxpr, Literal, get_aval,
                     AbstractValue, unit, unitvar, abstract_unit, Primitive,
                     call_p, TypedJaxpr, new_jaxpr_eqn)
@@ -57,7 +56,7 @@ class JaxprTrace(Trace):
     return JaxprTracer(self, val.pval, FreeVar(val))
 
   def new_const(self, val):
-    if isinstance(val, Tracer) and val.trace.level == self.level:
+    if isinstance(val, Tracer) and val._trace.level == self.level:
       raise Exception
     return JaxprTracer(self, PartialVal((None, val)), unit)
 
@@ -97,45 +96,56 @@ class JaxprTrace(Trace):
     if primitive in custom_partial_eval_rules:
       return custom_partial_eval_rules[primitive](self, *tracers, **params)
     else:
-      pvs, consts = unzip2(t.pval for t in tracers)
-      if all(pv is None for pv in pvs):
-        return primitive.bind(*consts, **params)
-      tracers = map(self.instantiate_const, tracers)
-      avals = [t.aval for t in tracers]
-      out_aval = primitive.abstract_eval(*avals, **params)
-      if primitive.multiple_results:
-        out_tracers = [JaxprTracer(self, PartialVal((aval, unit)), None)
-                       for aval in out_aval]
-        eqn = new_eqn_recipe(tracers, out_tracers, primitive, (), params)
-        for t in out_tracers: t.recipe = eqn
-        return out_tracers
-      else:
-        out_tracer = JaxprTracer(self, PartialVal((out_aval, unit)), None)
-        out_tracer.recipe = new_eqn_recipe(tracers, [out_tracer], primitive, (), params)
-        return out_tracer
+      return self.default_process_primitive(primitive, tracers, params)
 
-  def process_call(self, call_primitive, f, tracers, params):
+  def default_process_primitive(self, primitive, tracers, params):
+    pvs, consts = unzip2(t.pval for t in tracers)
+    if all(pv is None for pv in pvs):
+      return primitive.bind(*consts, **params)
+    tracers = map(self.instantiate_const, tracers)
+    avals = [t.aval for t in tracers]
+    out_aval = primitive.abstract_eval(*avals, **params)
+    if primitive.multiple_results:
+      out_tracers = [JaxprTracer(self, PartialVal((aval, unit)), None)
+                     for aval in out_aval]
+      eqn = new_eqn_recipe(tracers, out_tracers, primitive, params)
+      for t in out_tracers: t.recipe = eqn
+      return out_tracers
+    else:
+      out_tracer = JaxprTracer(self, PartialVal((out_aval, unit)), None)
+      out_tracer.recipe = new_eqn_recipe(tracers, [out_tracer], primitive, params)
+      return out_tracer
+
+  def process_call(self, call_primitive, f: lu.WrappedFun, tracers, params):
+    name = params.get('name', f.__name__)
     if self.master.trace_type is StagingJaxprTrace:
       tracers = map(self.instantiate_const_abstracted, tracers)
+    else:
+      name = wrap_name(name, 'pe')
+    params = dict(params, name=name)
     if call_primitive in call_partial_eval_rules:
-      return call_partial_eval_rules[call_primitive](self, f, tracers, params)
+      return call_partial_eval_rules[call_primitive](self, call_primitive, f, tracers, params)
     if call_primitive in map_primitives:
       return self.process_map(call_primitive, f, tracers, params)
     in_pvs, in_consts = unzip2([t.pval for t in tracers])
     fun, aux = partial_eval(f, self, in_pvs)
     out_flat = call_primitive.bind(fun, *in_consts, **params)
     out_pvs, jaxpr, env = aux()
+    env_tracers = map(self.full_raise, env)
     out_pv_consts, consts = split_list(out_flat, [len(out_flat)-len(jaxpr.constvars)])
     const_tracers = map(self.new_instantiated_const, consts)
-    bound_subjaxpr = (jaxpr, const_tracers, map(self.full_raise, env))
+    lifted_jaxpr = convert_constvars_jaxpr(jaxpr)
     out_tracers = [JaxprTracer(self, PartialVal((out_pv, out_pv_const)), None)
                    for out_pv, out_pv_const in zip(out_pvs, out_pv_consts)]
-    eqn = new_eqn_recipe(tracers, out_tracers, call_primitive, (bound_subjaxpr,), params)
+    new_params = dict(params, call_jaxpr=lifted_jaxpr)
+    # The `jaxpr` already contains the env_vars at start of invars
+    eqn = new_eqn_recipe(tuple(it.chain(const_tracers, env_tracers, tracers)),
+                         out_tracers, call_primitive, new_params)
     for t in out_tracers:
       t.recipe = eqn
     return out_tracers
 
-  def process_map(self, map_primitive, f, tracers, params):
+  def process_map(self, map_primitive, f: lu.WrappedFun, tracers, params):
     in_pvs, in_consts = unzip2([t.pval for t in tracers])
     reduced_pvs = [None if pv is None else _mapped_aval(pv) for pv in in_pvs]
     fun, aux = partial_eval(f, self, reduced_pvs)
@@ -145,12 +155,18 @@ class JaxprTrace(Trace):
     out_pvs = [None if pv is None else _unmapped_aval(params['axis_size'], pv)
                for pv in out_pvs_reduced]
     const_tracers = map(self.new_instantiated_const, consts)
-    lifted_jaxpr = closure_convert_jaxpr(jaxpr)
-    bound_subjaxpr = (lifted_jaxpr, (), map(self.full_raise, env))
+    env_tracers = map(self.full_raise, env)
+    lifted_jaxpr = convert_constvars_jaxpr(jaxpr)
     out_tracers = [JaxprTracer(self, PartialVal((out_pv, out_pv_const)), None)
                    for out_pv, out_pv_const in zip(out_pvs, out_pv_consts)]
-    eqn = new_eqn_recipe(tuple(it.chain(const_tracers, tracers)),
-                         out_tracers, map_primitive, (bound_subjaxpr,), params)
+    # The `jaxpr` already contains the env_vars at start of invars
+    new_params = dict(params,
+                      mapped_invars=tuple([True] * len(const_tracers) +
+                                          [False] * len(env_tracers) +
+                                          [True] * len(tracers)),
+                      call_jaxpr=lifted_jaxpr)
+    eqn = new_eqn_recipe(tuple(it.chain(const_tracers, env_tracers, tracers)),
+                         out_tracers, map_primitive, new_params)
     for t in out_tracers:
       t.recipe = eqn
     return out_tracers
@@ -169,10 +185,13 @@ class JaxprTrace(Trace):
       trace = JaxprTrace(master, core.cur_sublevel())
       const_tracers = map(trace.new_instantiated_const, consts)
       env_tracers = map(trace.full_raise, env)
-      bound_subjaxpr = (jaxpr, const_tracers, env_tracers)
+      lifted_jaxpr = convert_constvars_jaxpr(jaxpr)
       out_tracers = [JaxprTracer(trace, PartialVal((out_pv, out_pv_const)), None)
                      for out_pv, out_pv_const in zip(out_pvs, out_pv_consts)]
-      eqn = new_eqn_recipe([], out_tracers, call_primitive, (bound_subjaxpr,), params)
+      new_params = dict(params, call_jaxpr=lifted_jaxpr)
+      # The `jaxpr` already contains the env_vars at start of invars
+      eqn = new_eqn_recipe(tuple(it.chain(const_tracers, env_tracers)),
+                           out_tracers, call_primitive, new_params)
       for t in out_tracers:
         t.recipe = eqn
       return out_tracers
@@ -191,17 +210,39 @@ class JaxprTrace(Trace):
       out_pv_consts, consts = x[:n], x[n:]
       trace = JaxprTrace(master, core.cur_sublevel())
       const_tracers = map(trace.new_instantiated_const, consts)
-      env_tracers = map(trace.full_raise, env)
-      lifted_jaxpr = closure_convert_jaxpr(jaxpr)
-      bound_subjaxpr = (lifted_jaxpr, (), env_tracers)
+      # The `jaxpr` already contains the env_vars at start of invars
+      lifted_jaxpr = convert_constvars_jaxpr(jaxpr)
       out_tracers = [JaxprTracer(trace, PartialVal((out_pv, out_pv_const)), None)
                      for out_pv, out_pv_const in zip(out_pvs, out_pv_consts)]
-      eqn = new_eqn_recipe(const_tracers, out_tracers, map_primitive,
-                           (bound_subjaxpr,), params)
+      new_params = dict(params,
+                        mapped_invars=tuple([True] * len(const_tracers) +
+                                            [False] * len(env)),
+                        call_jaxpr=lifted_jaxpr)
+      env_tracers = map(trace.full_raise, env)
+      eqn = new_eqn_recipe(it.chain(const_tracers, env_tracers),
+                           out_tracers, map_primitive, new_params)
       for t in out_tracers:
         t.recipe = eqn
       return out_tracers
     return out, todo
+
+  def process_custom_jvp_call(self, prim, fun, jvp, tracers):
+    # We form jaxprs using JaxprTraces for two distinct purposes: to stage
+    # program representations completely out of the JAX system (e.g. for XLA
+    # using jit or pmap), and to build a representation of a function that may
+    # require further JAX transformations (e.g. in "initial-style" higher-order
+    # primitives, like for control flow). In particular, in the latter case we
+    # need custom differentiation rules to stick around, but in the former we do
+    # not. This method call should only be reachable in the former case, and so
+    # we check that the former case is indicated (with a StagingJaxprTrace) and
+    # then drop the differentiation rules.
+    assert self.master.trace_type is StagingJaxprTrace
+    return fun.call_wrapped(*tracers)
+
+  def process_custom_vjp_call(self, prim, fun, fwd, bwd, tracers, out_trees):
+    # See comment in the above process_custom_jvp_call method.
+    assert self.master.trace_type is StagingJaxprTrace
+    return fun.call_wrapped(*tracers)
 
 # This subclass is used just for its type tag, which switches the behavior of
 # process_call to stage out into the jaxpr any call primitives encountered
@@ -226,13 +267,13 @@ def _unmapped_aval(size, aval):
   else:
     raise TypeError(aval)
 
-map_primitives = set()
-custom_partial_eval_rules = {}
-call_partial_eval_rules = {}
+map_primitives: Set[core.Primitive] = set()
+custom_partial_eval_rules: Dict[core.Primitive, Callable] = {}
+call_partial_eval_rules: Dict[core.Primitive, Callable] = {}
 
 
-def partial_eval(f, trace, pvs):
-  f = trace_to_subjaxpr(f, trace.master, False)
+def partial_eval(f, trace, pvs, instantiate=False):
+  f = trace_to_subjaxpr(f, trace.master, instantiate)
   return partial_eval_wrapper(f, tuple(pvs))
 
 
@@ -248,7 +289,7 @@ def partial_eval_wrapper(avals, *consts):
 def abstract_eval_fun(fun, *avals, **params):
   pvals_in = [PartialVal((a, unit)) for a in avals]
   _, pvals_out, _ = trace_to_jaxpr(lu.wrap_init(fun, params), pvals_in,
-                                  instantiate=True)
+                                  instantiate=True, stage_out=True)
   avals_out, _ = unzip2(pvals_out)
   for aval_out in avals_out:
     assert isinstance(aval_out, AbstractValue)  # instantiate=True
@@ -261,14 +302,15 @@ class JaxprTracer(Tracer):
   def __init__(self, trace, pval, recipe):
     assert isinstance(pval, PartialVal)
     pv, const = pval
-    if isinstance(const, Tracer):
-      assert const.trace.level < trace.level
-    self.trace = trace
+    if isinstance(const, Tracer) and const._trace.level >= trace.level:
+      raise core.escaped_tracer_error(
+          "Tracer from a higher level: {} in trace {}".format(const, trace))
+    self._trace = trace
     self.pval = pval
     self.recipe = recipe
 
   def __repr__(self):
-    return 'Traced<{}:{}>'.format(self.aval, self.trace)
+    return 'Traced<{}:{}>'.format(self.aval, self._trace)
 
   @property
   def aval(self):
@@ -278,7 +320,7 @@ class JaxprTracer(Tracer):
   @property
   def parents(self):
     if isinstance(self.recipe, JaxprEqnRecipe):
-      return eqn_parents(self.recipe)
+      return self.recipe.invars
     else:
       return []
 
@@ -307,7 +349,7 @@ class PartialVal(tuple):
 
 valid_pv_types = (AbstractValue, type(None))
 
-def merge_pvals(val, pval):
+def merge_pvals(val, pval: PartialVal):
   pv, const = pval
   if isinstance(pv, AbstractValue):
     return val
@@ -324,10 +366,10 @@ def partial_val_aval(pv, const):
   else:
     raise TypeError(pv)
 
-def trace_to_jaxpr(fun, pvals, instantiate=False, stage_out_calls=False):
-  """Traces a function, given abstract inputs, to a jaxpr."""
-  trace_type = StagingJaxprTrace if stage_out_calls else JaxprTrace
-  with new_master(trace_type) as master:
+def trace_to_jaxpr(fun: lu.WrappedFun, pvals: Sequence[PartialVal],
+                   instantiate=False, stage_out=False, bottom=False):
+  trace_type = StagingJaxprTrace if stage_out else JaxprTrace
+  with new_master(trace_type, bottom=bottom) as master:
     fun = trace_to_subjaxpr(fun, master, instantiate)
     jaxpr, (out_pvals, consts, env) = fun.call_wrapped(pvals)
     assert not env
@@ -361,47 +403,80 @@ FreeVar = namedtuple('FreeVar', ['val'])
 ConstVar = namedtuple('ConstVar', ['val'])
 LambdaBinding = namedtuple('LambdaBinding', [])
 JaxprEqnRecipe = namedtuple('JaxprEqnRecipe',
-                            ['eqn_id', 'invars', 'outvars', 'primitive',
-                             'bound_subjaxprs', 'params'])
+                            ['eqn_id', 'invars', 'outvars', 'primitive', 'params'])
 
-def new_eqn_recipe(invars, outvars, primitive, bound_subjaxprs, params):
-  return JaxprEqnRecipe(object(), invars, map(ref, outvars), primitive,
-                        bound_subjaxprs, params)
+def new_eqn_recipe(invars, outvars, primitive, params):
+  """Constructs a new JaxEqnRecipe.
+
+  Params:
+    invars: the tracers for the primitive inputs.
+    outvars: the tracers for the primitive outputs.
+    primitive: the primitive.
+    params: the primitive params
+  """
+  if primitive.call_primitive:
+    # TODO(necula): move these checks to core.check_jaxpr, and call it
+    # in more places.
+    assert "call_jaxpr" in params
+  return JaxprEqnRecipe(object(), tuple(invars), map(ref, outvars), primitive,
+                        params)
+
 
 def recipe_to_eqn(unused_var, getvar, recipe):
-  _, in_tracers, out_tracer_refs, primitive, bound_subjaxprs, params = recipe
+  _, in_tracers, out_tracer_refs, primitive, params = recipe
   out_tracers = [t_ref() for t_ref in out_tracer_refs]
   invars  = [getvar(t) for t in in_tracers]
   outvars = [unused_var() if t is None else getvar(t) for t in out_tracers]
-  new_bound_subjaxprs = [(j, map(getvar, c), map(getvar, f))
-                         for j, c, f in bound_subjaxprs]
-  return new_jaxpr_eqn(invars, outvars, primitive, new_bound_subjaxprs, params)
+  return new_jaxpr_eqn(invars, outvars, primitive, params)
 
 def tracers_to_jaxpr(in_tracers, out_tracers):
+  """Constructs Jaxpr given tracers for inputs and outputs.
+
+  Params:
+    in_tracers: the tracers that were created for the function inputs
+    out_tracers: the tracers that were output by the function.
+
+  Returns: a triple of a `Jaxpr`, a list of constant values corresponding to
+    the `constvars` in the returned Jaxps, and a list of environment values.
+    The vars for the environment values have been pre-pended to the Jaxpr's
+    `invars`.
+  """
   newvar = core.gensym('')
-  t_to_var = defaultdict(newvar)
-  getvar = lambda t: t_to_var[id(t)]
+  t_to_var = {}
+  def getvar(t):
+    var = t_to_var.get(id(t))
+    if var is None:
+      var = newvar(partial_val_aval(*t.pval))
+      t_to_var[id(t)] = var
+    return var
   sorted_tracers = toposort(out_tracers)
   invars = map(getvar, in_tracers)
   eqns = []
   env = {}
   consts = {}
-  const_to_var = defaultdict(newvar)
-  destructuring_vars = {}
+  const_to_var = {}
+  def getconstvar(c):
+    var = const_to_var.get(id(c))
+    if var is None:
+      var = newvar(get_aval(c))
+      const_to_var[id(c)] = var
+    return var
   processed_eqn_ids = set()
   for t in sorted_tracers:
     recipe = t.recipe
     if isinstance(recipe, JaxprEqnRecipe):
       if recipe.eqn_id not in processed_eqn_ids:
-        eqns.append(recipe_to_eqn(newvar, getvar, recipe))
+        eqns.append(recipe_to_eqn(lambda: newvar(core.abstract_unit), getvar, recipe))
         processed_eqn_ids.add(recipe.eqn_id)
     elif isinstance(recipe, LambdaBinding):
-      assert any(t is in_tracer for in_tracer in in_tracers), "Encountered unexpected tracer"
+      if not any(t is in_tracer for in_tracer in in_tracers):
+        raise core.escaped_tracer_error(
+            "Tracer not among input tracers {}".format(t))
       assert in_tracers, "Lambda binding with no args"
     elif isinstance(recipe, FreeVar):
       env[getvar(t)] = recipe.val
     elif isinstance(recipe, ConstVar):
-      v = t_to_var[id(t)] = const_to_var[id(recipe.val)]
+      v = t_to_var[id(t)] = getconstvar(recipe.val)
       consts[v] = recipe.val
     elif isinstance(recipe, Literal):
       t_to_var[id(t)] = recipe
@@ -412,28 +487,17 @@ def tracers_to_jaxpr(in_tracers, out_tracers):
 
   env_vars, env_vals = unzip2(env.items())
   const_vars, const_vals = unzip2(consts.items())
-  jaxpr = Jaxpr(const_vars, env_vars, invars, list(map(getvar, out_tracers)), eqns)
+  # The env_vars are pre-pended to the invars
+  jaxpr = Jaxpr(const_vars, list(it.chain(env_vars, invars)), list(map(getvar, out_tracers)), eqns)
   core.skip_checks or core.check_jaxpr(jaxpr)
   return jaxpr, const_vals, env_vals
 
-
-
-def eqn_parents(eqn):
-  subjaxpr_tracers = [it.chain(c, f) for _, c, f in eqn.bound_subjaxprs]
-  return list(it.chain(eqn.invars,  *subjaxpr_tracers))
-
-def closure_convert_jaxpr(jaxpr):
+@cache()
+def convert_constvars_jaxpr(jaxpr):
+  """Moves the constvars to the start of invars."""
   core.skip_checks or core.check_jaxpr(jaxpr)
-  lifted_jaxpr = Jaxpr(constvars=(), freevars=jaxpr.freevars,
+  lifted_jaxpr = Jaxpr(constvars=(),
                        invars=jaxpr.constvars + jaxpr.invars,
-                       outvars=jaxpr.outvars, eqns=jaxpr.eqns)
-  core.skip_checks or core.check_jaxpr(lifted_jaxpr)
-  return lifted_jaxpr
-
-def convert_freevars_jaxpr(jaxpr):
-  core.skip_checks or core.check_jaxpr(jaxpr)
-  lifted_jaxpr = Jaxpr(constvars=jaxpr.constvars, freevars=(),
-                       invars=jaxpr.freevars + jaxpr.invars,
                        outvars=jaxpr.outvars, eqns=jaxpr.eqns)
   core.skip_checks or core.check_jaxpr(lifted_jaxpr)
   return lifted_jaxpr
@@ -460,8 +524,12 @@ def partial_eval_jaxpr(jaxpr, unknowns, instantiate):
   # jaxpr_1 :: a1 -> [b1, res]
   # jaxpr_2 :: res | a2 -> b2
   # jaxpr_2 :: [a2, res] -> b2
-  jaxpr_2 = closure_convert_jaxpr(jaxpr_2)
+  jaxpr_2 = convert_constvars_jaxpr(jaxpr_2)
   jaxpr_2.invars = jaxpr_2.invars[num_res:] + jaxpr_2.invars[:num_res]
+  for var, unknown in zip(jaxpr_2.invars[:len(unknowns)], unknowns):
+    if not unknown:
+      var.aval = abstract_unit
+
   uk_out = [pv is not None for pv in out_pvs_2]
 
   in_avals_1, in_avals_2 = unzip2(map(_split_aval, unknowns, jaxpr.in_avals))
@@ -482,12 +550,13 @@ def _split_aval(unknown, aval):
 
 
 remat_call_p = core.Primitive('remat_call')
+remat_call_p.call_primitive = True
 remat_call = partial(core.call_bind, remat_call_p)
 remat_call_p.def_custom_bind(remat_call)
 remat_call_p.def_impl(core.call_impl)
 remat_call_p.multiple_results = True
 
-def _remat_partial_eval(trace, f, tracers, params):
+def _remat_partial_eval(trace, _, f, tracers, params):
   concrete = params['concrete']
 
   # Unlike JaxprTrace.process_call, we want to form a jaxpr for the entirety of
@@ -507,11 +576,7 @@ def _remat_partial_eval(trace, f, tracers, params):
   # Using the instantiated tracers, run call_bind like JaxprTrace.process_call.
   in_pvs, in_consts = unzip2(t.pval for t in instantiated_tracers)
   fun, aux = partial_eval(f, trace, in_pvs)
-  if concrete:
-    # TODO(mattjj): remove `remat_context` when confident no accidental FLOPs
-    with remat_context():
-      out_flat = remat_call_p.bind(fun, *in_consts, **params)
-  else:
+  with core.initial_style_staging():
     out_flat = remat_call_p.bind(fun, *in_consts, **params)
   out_pvs, jaxpr, env = aux()
   env = map(trace.full_raise, env)
@@ -520,15 +585,15 @@ def _remat_partial_eval(trace, f, tracers, params):
 
   # Since we traced with everything marked as unknown, but we need to know which
   # outputs are known/unknown, we use partial_eval_jaxpr to get out_unknowns.
-  jaxpr_converted = convert_freevars_jaxpr(jaxpr)
-  in_avals = ([raise_to_shaped(t.pval[0]) for t in env]
+
+  in_avals = ([raise_to_shaped(partial_val_aval(*t.pval)) for t in env]
               + [raise_to_shaped(pv) for pv in in_pvs])
   out_avals = [raise_to_shaped(pv if pv is not None
                                else abstract_unit if var is unitvar
                                else get_aval(var.val) if type(var) is Literal
                                else get_aval(const))
                for var, pv, const in zip(jaxpr.outvars, out_pvs, out_pval_consts1)]
-  typed_jaxpr = core.TypedJaxpr(jaxpr_converted, consts, in_avals, out_avals)
+  typed_jaxpr = core.TypedJaxpr(jaxpr, consts, in_avals, out_avals)
   in_unknowns = [t.pval[0] is not None for t in it.chain(env, tracers)]
   jaxpr_1, jaxpr_2, out_unknowns = partial_eval_jaxpr(typed_jaxpr, in_unknowns, False)
   num_res = len(jaxpr_1.out_avals) - len(jaxpr_2.out_avals)
@@ -553,10 +618,11 @@ def _remat_partial_eval(trace, f, tracers, params):
   # Now that we have out_pvals, the rest is just like JaxprTrace.process_call.
   instantiated_tracers = env + instantiated_tracers
   const_tracers = map(trace.new_instantiated_const, consts)
-  bound_subjaxpr = (typed_jaxpr.jaxpr, const_tracers, ())
+  lifted_jaxpr = convert_constvars_jaxpr(typed_jaxpr.jaxpr)
   out_tracers = [JaxprTracer(trace, out_pval, None) for out_pval in out_pvals]
-  eqn = new_eqn_recipe(instantiated_tracers, out_tracers, remat_call_p,
-                       (bound_subjaxpr,), params)
+  new_params = dict(params, call_jaxpr=lifted_jaxpr)
+  eqn = new_eqn_recipe(tuple(it.chain(const_tracers, instantiated_tracers)),
+                       out_tracers, remat_call_p, new_params)
   for t in out_tracers: t.recipe = eqn
   return out_tracers
 call_partial_eval_rules[remat_call_p] = _remat_partial_eval
@@ -578,8 +644,7 @@ def _dce_jaxpr(typed_jaxpr, outputs):
       new_eqns.append(eqn)
       needed_vars.update(v for v in eqn.invars if type(v) is not Literal)
   new_eqns = new_eqns[::-1]
-
-  new_jaxpr = core.Jaxpr(jaxpr.constvars, jaxpr.freevars, jaxpr.invars,
+  new_jaxpr = core.Jaxpr(jaxpr.constvars, jaxpr.invars,
                          new_outvars, new_eqns)
   return core.TypedJaxpr(new_jaxpr, typed_jaxpr.literals, typed_jaxpr.in_avals,
                          new_out_avals)
@@ -594,29 +659,12 @@ def _reconstruct_pval(pval1, const2, unknown):
     else:
       return PartialVal((None, const2))
 
-# TODO(mattjj): for https://github.com/google/jax/pull/1749 we allowed
-# standard_abstract_eval to perform concrete evaluation (i.e. FLOPs), but we
-# don't think it should happen except for in a remat context
-@contextlib.contextmanager
-def remat_context():
-  try:
-    prev_state = _thread_local_state.remat
-    _thread_local_state.remat = True
-    yield
-  finally:
-    _thread_local_state.remat = prev_state
-
-class _ThreadLocalState(threading.local):
-  def __init__(self):
-    self.remat = False
-_thread_local_state = _ThreadLocalState()
-
 
 def move_binders_to_front(typed_jaxpr, to_move):
-  assert not typed_jaxpr.jaxpr.constvars and not typed_jaxpr.jaxpr.freevars
+  assert not typed_jaxpr.jaxpr.constvars
   assert len(typed_jaxpr.in_avals) == len(to_move)
   new_invars = _move_to_front(typed_jaxpr.jaxpr.invars, to_move)
-  new_jaxpr = core.Jaxpr((), (), new_invars, typed_jaxpr.jaxpr.outvars,
+  new_jaxpr = core.Jaxpr((), new_invars, typed_jaxpr.jaxpr.outvars,
                          typed_jaxpr.jaxpr.eqns)
   new_in_avals = _move_to_front(typed_jaxpr.in_avals, to_move)
   new_typed_jaxpr = core.TypedJaxpr(new_jaxpr, typed_jaxpr.literals,
@@ -626,4 +674,3 @@ def move_binders_to_front(typed_jaxpr, to_move):
 def _move_to_front(lst, to_move):
   return ([elt for elt, move in zip(lst, to_move) if move] +
           [elt for elt, move in zip(lst, to_move) if not move])
-
