@@ -12,15 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
 
-from collections import namedtuple, defaultdict
+from collections import defaultdict
 from contextlib import contextmanager
-import itertools as it
 import operator as op
 import threading
+from typing import Any, Callable, Dict, List, Sequence, Set, Type
 
 from absl import logging
 import numpy as onp
@@ -31,9 +28,9 @@ from .. import linear_util as lu
 from .. import lazy
 from ..abstract_arrays import (ConcreteArray, ShapedArray, array_types,
                                raise_to_shaped)
-from ..util import partial, unzip2, concatenate, prod, safe_map
+from ..util import (partial, unzip2, prod, safe_map,
+                    extend_name_stack, wrap_name)
 from ..lib import xla_bridge as xb
-from .xla import aval_to_xla_shape, xla_destructure
 from .batching import broadcast, not_mapped
 from . import batching
 from . import partial_eval as pe
@@ -49,7 +46,7 @@ _map = safe_map
 
 def identity(x): return x
 
-def shard_args(backend, devices, assignments, axis_size, tuple_args, args):
+def shard_args(backend, devices, assignments, axis_size, args):
   """Shard each argument data array along its leading axis.
 
   Args:
@@ -100,14 +97,10 @@ def shard_args(backend, devices, assignments, axis_size, tuple_args, args):
       for r, buf in enumerate(bufs):
         buffers[r][a] = buf
 
-  if tuple_args:
-    buffers = [[xla.make_tuple(bufs, devices[r], backend)]
-               for r, bufs in enumerate(buffers)]
-
   return buffers
 
 
-shard_arg_handlers = {}
+shard_arg_handlers: Dict[Any, Callable[[Any, Any, Any], Sequence[Any]]] = {}
 shard_arg_handlers[core.Unit] = \
     lambda x, devices, _: [xla.device_put(core.unit, d) for d in devices]
 def _shard_array(x, devices, assignments):
@@ -126,9 +119,10 @@ shard_arg_handlers[xla.DeviceArray] = _shard_device_array
 def shard_aval(size, aval):
   try:
     return shard_aval_handlers[type(aval)](size, aval)
-  except KeyError:
-    raise TypeError("No shard_aval handler for type: {}".format(type(aval)))
-shard_aval_handlers = {}
+  except KeyError as err:
+    raise TypeError("No shard_aval handler for type: {}".format(type(aval))
+                    ) from err
+shard_aval_handlers: Dict[Type[core.AbstractValue], Callable[[int, Any], Any]] = {}
 shard_aval_handlers[core.AbstractUnit] = lambda size, x: x
 def _shard_abstract_array(size, x):
   if x.shape[0] != size:
@@ -140,9 +134,11 @@ shard_aval_handlers[ShapedArray] = _shard_abstract_array
 def aval_to_result_handler(size, nrep, aval):
   try:
     return pxla_result_handlers[type(aval)](size, nrep, aval)
-  except KeyError:
-    raise TypeError("No pxla_result_handler for type: {}".format(type(aval)))
-pxla_result_handlers = {}
+  except KeyError as err:
+    raise TypeError("No pxla_result_handler for type: {}".format(type(aval))
+                    ) from err
+PxlaResultHandler = Callable[..., Callable[[Any], Any]]
+pxla_result_handlers: Dict[Type[core.AbstractValue], PxlaResultHandler] = {}
 pxla_result_handlers[core.AbstractUnit] = lambda *_: lambda _: core.unit
 def array_result_handler(size, nrep, aval):
   full_aval = ShapedArray((size,) + aval.shape, aval.dtype)
@@ -160,7 +156,7 @@ def assign_shards_to_replicas(nrep, size):
 
   Returns:
     A tuple of integers of length nrep in which the elements take on values from
-    0 to size-1. Replica n is assgined shard data_array[assignments[n]].
+    0 to size-1. Replica n is assigned shard data_array[assignments[n]].
   """
   groupsize, ragged = divmod(nrep, size)
   assert not ragged
@@ -248,10 +244,47 @@ def apply_parallel_primitive(prim, *args, **params):
     shape = (logical_size(dynamic_axis_env[axis_name]),)
   return parallel_pure_rules[prim](*args, shape=shape, **params)
 
-parallel_pure_rules = {}
+parallel_pure_rules: Dict[core.Primitive, Callable] = {}
 
 
 def axis_index(axis_name):
+  """Return the index along the pmapped axis ``axis_name``.
+
+  Args:
+    axis_name: hashable Python object used to name the pmapped axis (see the
+      ``pmap`` docstring for more details).
+
+  Returns:
+    An integer representing the index.
+
+  For example, with 8 XLA devices available:
+
+  >>> from functools import partial
+  >>> @partial(pmap, axis_name='i')
+  ... def f(_):
+  ...   return lax.axis_index('i')
+  ...
+  >>> f(np.zeros(4))
+  ShardedDeviceArray([0, 1, 2, 3], dtype=int32)
+  >>> f(np.zeros(8))
+  ShardedDeviceArray([0, 1, 2, 3, 4, 5, 6, 7], dtype=int32)
+  >>> @partial(pmap, axis_name='i')
+  ... @partial(pmap, axis_name='j')
+  ... def f(_):
+  ...   return lax.axis_index('i'), lax.axis_index('j')
+  ...
+  >>> x, y = f(np.zeros((4, 2)))
+  >>> print(x)
+  [[0 0]
+   [1 1]
+   [2 2]
+   [3 3]]
+  >>> print(y)
+  [[0 1]
+   [0 1]
+   [0 1]
+   [0 1]]
+  """
   dynamic_axis_env = _thread_local_state.dynamic_axis_env
   frame = dynamic_axis_env[axis_name]
   sizes = dynamic_axis_env.sizes[:dynamic_axis_env.index(frame)+1]
@@ -269,7 +302,7 @@ def _axis_index_partial_eval(trace, _, **params):
   # rule except that we don't attempt to lower out of the trace.
   out_aval = ShapedArray((), onp.int32)
   out_tracer = pe.JaxprTracer(trace, pe.PartialVal((out_aval, core.unit)), None)
-  eqn = pe.new_eqn_recipe([], [out_tracer], axis_index_p, (), params)
+  eqn = pe.new_eqn_recipe([], [out_tracer], axis_index_p, params)
   out_tracer.recipe = eqn
   return out_tracer
 
@@ -383,7 +416,7 @@ xla.canonicalize_dtype_handlers[ShardedDeviceArray] = identity
 
 
 class ChunkedDeviceArray(ShardedDeviceArray):
-  __slots__ = []
+  __slots__: List[str] = []
   _collect = staticmethod(onp.concatenate)
 
   def __init__(self, axis_size, aval, device_buffers):
@@ -403,22 +436,16 @@ xla.canonicalize_dtype_handlers[ChunkedDeviceArray] = identity
 
 ### the xla_pmap primitive and its rules are comparable to xla_call in xla.py
 
-def xla_pmap_impl(fun, *args, **params):
-  axis_name = params.pop('axis_name')
-  axis_size = params.pop('axis_size')
-  global_axis_size = params.pop('global_axis_size')
-  devices = params.pop('devices')
-  backend = params.pop('backend')
-  assert not params
-
+def xla_pmap_impl(fun: lu.WrappedFun, *args, backend, axis_name, axis_size, global_axis_size,
+                  devices, name, mapped_invars=None):
   abstract_args = map(xla.abstractify, args)
   compiled_fun = parallel_callable(fun, backend, axis_name, axis_size,
-                                   global_axis_size, devices, *abstract_args)
+                                   global_axis_size, devices, name, *abstract_args)
   return compiled_fun(*args)
 
 @lu.cache
 def parallel_callable(fun, backend, axis_name, axis_size, global_axis_size,
-                      devices, *avals):
+                      devices, name, *avals):
   if devices is not None and len(devices) == 0:
     raise ValueError("'devices' argument to pmap must be non-empty, or None.")
 
@@ -461,18 +488,17 @@ def parallel_callable(fun, backend, axis_name, axis_size, global_axis_size,
 
   @lu.wrap_init
   def dynamic_fun(dummy, *args):
-    with extend_dynamic_axis_env(axis_name, dummy.trace, global_axis_size):
+    with extend_dynamic_axis_env(axis_name, dummy._trace, global_axis_size):
       return fun.call_wrapped(*args)
 
   avals = tuple(map(partial(shard_aval, axis_size), avals))
   pvals = [pe.PartialVal((aval, core.unit)) for aval in avals]
+  # We add a dummy first invar, to carry the trace details to `dynamic_fun`
   pval = pe.PartialVal([core.abstract_unit, core.unit])  # dummy value for axis env
-  with core.new_master(pe.StagingJaxprTrace, True) as master:
-    jaxpr, (out_pvals, consts, env) = pe.trace_to_subjaxpr(
-        dynamic_fun, master, False).call_wrapped([pval] + pvals)
-    jaxpr.invars = jaxpr.invars[1:]  # ignore dummy
-    assert not env
-    del master
+  jaxpr, out_pvals, consts = pe.trace_to_jaxpr(
+      dynamic_fun, [pval] + pvals, instantiate=False, stage_out=True, bottom=True)
+  jaxpr.invars = jaxpr.invars[1:]  # ignore dummy
+
   out_pvs, out_consts = unzip2(out_pvals)
 
   # TODO(skye,mattjj): allow more collectives on multi-host as we test them, but
@@ -501,14 +527,15 @@ def parallel_callable(fun, backend, axis_name, axis_size, global_axis_size,
   jaxpr_replicas = xla.jaxpr_replicas(jaxpr)
   num_local_replicas = axis_size * jaxpr_replicas
   num_global_replicas = global_axis_size * jaxpr_replicas
-  axis_env = xla.AxisEnv(num_global_replicas, [axis_name], [global_axis_size], devices)
+  axis_env = xla.AxisEnv(num_global_replicas, (axis_name,), (global_axis_size,), devices)
 
   tuple_args = len(avals) > 100  # pass long arg lists as tuple for TPU
 
   c = xb.make_computation_builder("pmap_{}".format(fun.__name__))
   xla_consts = _map(c.Constant, consts)
   xla_args = xla._xla_callable_args(c, avals, tuple_args)
-  out_nodes = xla.jaxpr_subcomp(c, jaxpr, backend, axis_env, xla_consts, (), *xla_args)
+  out_nodes = xla.jaxpr_subcomp(c, jaxpr, backend, axis_env, xla_consts,
+                                extend_name_stack(wrap_name(name, 'pmap')), *xla_args)
   built = c.Build(c.Tuple(*out_nodes))
 
   if devices is None:
@@ -546,18 +573,22 @@ def parallel_callable(fun, backend, axis_name, axis_size, global_axis_size,
 
   device_assignment = tuple(d.id for d in devices)
   compiled = built.Compile(
-      compile_options=xb.get_compile_options(num_global_replicas, device_assignment),
+      compile_options=xb.get_compile_options(
+          num_replicas=num_global_replicas,
+          num_partitions=1,
+          device_assignment=device_assignment),
       backend=xb.get_backend(backend))
 
   handle_args = partial(shard_args, backend, compiled.local_devices(),
                         assign_shards_to_replicas(num_local_replicas, axis_size),
-                        axis_size, tuple_args)
+                        axis_size)
   handle_outs = _pvals_to_results_handler(axis_size, num_local_replicas,
                                           out_pvals, compiled.local_devices(),
                                           backend)
-  return partial(execute_replicated, compiled, backend, handle_args, handle_outs)
+  return partial(execute_replicated, compiled, backend, handle_args, handle_outs,
+                 tuple_args)
 
-multi_host_supported_collectives = set()
+multi_host_supported_collectives: Set[core.Primitive] = set()
 
 class ResultToPopulate(object): pass
 result_to_populate = ResultToPopulate()
@@ -569,7 +600,7 @@ def _pvals_to_results_handler(size, nrep, out_pvals, devices, backend):
   def handler(out_bufs):
     buffers = [[result_to_populate] * nrep for _ in range(nouts)]
     for r, tuple_buf in enumerate(out_bufs):
-      for i, buf in enumerate(tuple_buf.destructure()):
+      for i, buf in enumerate(tuple_buf):
         buffers[i][r] = buf
     assert not any(buf is result_to_populate for bufs in buffers
                    for buf in bufs)
@@ -608,7 +639,7 @@ def replicate(val, axis_size, nrep, devices=None, backend=None):
     devices = xb.get_backend(backend).get_default_device_assignment(nrep)
   assert nrep == len(devices)
 
-  aval = xla.abstractify(val)
+  aval = xla.abstractify(val)  # type: ShapedArray
   aval = ShapedArray((axis_size,) + aval.shape, aval.dtype)
   device_buffers = [xla.device_put(val, d) for d in devices]
   return ShardedDeviceArray(aval, device_buffers)
@@ -636,21 +667,25 @@ def _pval_to_result_handler(axis_size, nrep, pval, devices, backend):
   else:
     return aval_to_result_handler(axis_size, nrep, pv)
 
-def execute_replicated(compiled, backend, in_handler, out_handler, *args):
+def execute_replicated(compiled, backend, in_handler, out_handler, tuple_args,
+                       *args):
   input_bufs = in_handler(args)
-  out_bufs = compiled.ExecutePerReplica(list(input_bufs))
+  out_bufs = compiled.ExecuteOnLocalDevices(
+      list(input_bufs), tuple_arguments=tuple_args)
   return out_handler(out_bufs)
 
 
 xla_pmap_p = core.Primitive('xla_pmap')
+xla_pmap_p.call_primitive = True
 xla_pmap_p.multiple_results = True
 xla_pmap = partial(core.call_bind, xla_pmap_p)
 xla_pmap_p.def_custom_bind(xla_pmap)
 xla_pmap_p.def_impl(xla_pmap_impl)
 
-def _pmap_translation_rule(c, jaxpr, axis_env, const_nodes, freevar_nodes,
-                           in_nodes, axis_name, axis_size, global_axis_size,
-                           devices, backend=None):
+def _pmap_translation_rule(c, axis_env,
+                           in_nodes, name_stack, axis_name, axis_size,
+                           global_axis_size, devices, name,
+                           call_jaxpr, *, backend=None, mapped_invars):
   # We in-line here rather than generating a Call HLO as in the xla_call
   # translation rule just because the extra tuple stuff is a pain.
   if axis_env.devices is not None or (axis_env.names and devices is not None):
@@ -658,41 +693,62 @@ def _pmap_translation_rule(c, jaxpr, axis_env, const_nodes, freevar_nodes,
   if global_axis_size is None:
     global_axis_size = axis_size
   new_env = xla.extend_axis_env(axis_env, axis_name, global_axis_size)
-  in_nodes_sharded = list(map(partial(_xla_shard, c, new_env), in_nodes))
-  sharded_outs = xla.jaxpr_subcomp(c, jaxpr, backend, new_env, const_nodes,
-                                   freevar_nodes, *in_nodes_sharded)
-  outs = [_xla_unshard(c, new_env, shard) for shard in sharded_outs]
+  # Shard the in_nodes that are mapped
+  in_avals = [v.aval for v in call_jaxpr.invars]
+  in_nodes_sharded = (
+    _xla_shard(c, aval, new_env, in_node) if in_node_mapped else in_node
+    for aval, in_node, in_node_mapped in zip(in_avals, in_nodes, mapped_invars))
+
+  sharded_outs = xla.jaxpr_subcomp(
+      c, call_jaxpr, backend, new_env, (),
+      extend_name_stack(name_stack, wrap_name(name, 'pmap')), *in_nodes_sharded)
+  out_avals = [v.aval for v in call_jaxpr.outvars]
+  outs = [_xla_unshard(c, aval, new_env, shard, backend=backend)
+          for aval, shard in zip(out_avals, sharded_outs)]
   return c.Tuple(*outs)
 
 xla.call_translations[xla_pmap_p] = _pmap_translation_rule
 ad.primitive_transposes[xla_pmap_p] = partial(ad.map_transpose, xla_pmap_p)
 pe.map_primitives.add(xla_pmap_p)
 
-def _xla_shard(c, axis_env, x):
-  xla_shape = c.GetShape(x)
-  if xla_shape.is_tuple():
-    assert not xla_shape.tuple_shapes()
+def _xla_shard(c, aval, axis_env, x):
+  if aval is core.abstract_unit:
     return x
-  else:
-    dims = list(xla_shape.dimensions())
+  elif isinstance(aval, ShapedArray):
+    dims = list(c.GetShape(x).dimensions())
     zero = c.Constant(onp.zeros((), dtype=onp.uint32))
     idxs = [_unravel_index(c, axis_env)] + [zero] * (len(dims) - 1)
     return c.Reshape(c.DynamicSlice(x, idxs, [1] + dims[1:]), None, dims[1:])
+  else:
+    raise TypeError((aval, c.GetShape(x)))
 
 # TODO(b/110096942): more efficient gather
-def _xla_unshard(c, axis_env, x):
-  xla_shape = c.GetShape(x)
-  if xla_shape.is_tuple():
-    assert not xla_shape.tuple_shapes()
+def _xla_unshard(c, aval, axis_env, x, backend):
+  if aval is core.abstract_unit:
     return x
-  else:
+  elif isinstance(aval, ShapedArray):
+    # TODO(mattjj): remove this logic when AllReduce PRED supported on CPU / GPU
+    convert_bool = (onp.issubdtype(aval.dtype, onp.bool_)
+                    and xb.get_backend(backend).platform in ('cpu', 'gpu'))
+    if convert_bool:
+      x = c.ConvertElementType(x, xb.dtype_to_etype(onp.float32))
+
+    xla_shape = c.GetShape(x)
     dims = list(xla_shape.dimensions())
     padded = c.Broadcast(c.Constant(onp.array(0, xla_shape.numpy_dtype())),
                          [axis_env.sizes[-1]] + dims)
     zero = c.Constant(onp.zeros((), dtype=onp.uint32))
     idxs = [_unravel_index(c, axis_env)] + [zero] * len(dims)
     padded = c.DynamicUpdateSlice(padded, c.Reshape(x, None, [1] + dims), idxs)
-    return c.CrossReplicaSum(padded, xla.axis_groups(axis_env, axis_env.names[-1]))
+    out = c.CrossReplicaSum(padded, xla.axis_groups(axis_env, axis_env.names[-1]))
+
+    # TODO(mattjj): remove this logic when AllReduce PRED supported on CPU / GPU
+    if convert_bool:
+      nonzero = c.Ne(out, c.Constant(onp.array(0, dtype=onp.float32)))
+      out = c.ConvertElementType(nonzero, xb.dtype_to_etype(onp.bool_))
+    return out
+  else:
+    raise TypeError((aval, c.GetShape(x)))
 
 def _unravel_index(c, axis_env):
   div = c.Constant(onp.array(axis_env.nreps // prod(axis_env.sizes), onp.uint32))
@@ -745,7 +801,7 @@ def add_chunk_to_axis_env(axis_name, soft_trace, soft_size):
 
 class SplitAxisTracer(core.Tracer):
   def __init__(self, trace, axis_name, val):
-    self.trace = trace
+    self._trace = trace
     self.axis_name = axis_name
     self.val = val
 
@@ -791,17 +847,23 @@ class SplitAxisTrace(core.Trace):
           # if the name matches this tracer's name, apply the split_axis rule
           try:
             rule = split_axis_rules[primitive]
-          except KeyError:
+          except KeyError as err:
             msg = "split_axis for {} not implemented. Open a feature request!"
-            raise NotImplementedError(msg.format(primitive))
+            raise NotImplementedError(msg.format(primitive)) from err
           which_mapped = [n is not not_mapped for n in names_in]
           val_out, is_mapped = rule(vals_in, which_mapped, **params)
           name_out = name if is_mapped else not_mapped
-          return SplitAxisTracer(self, name_out, val_out)
+          if primitive.multiple_results:
+            return [SplitAxisTracer(self, name_out, v) for v in val_out]
+          else:
+            return SplitAxisTracer(self, name_out, val_out)
         else:
           # if not, bind the primitive without any processing
           val_out = primitive.bind(*vals_in, **params)
-          return SplitAxisTracer(self, name, val_out)
+          if primitive.multiple_results:
+            return [SplitAxisTracer(self, name, v) for v in val_out]
+          else:
+            return SplitAxisTracer(self, name, val_out)
       else:
         # if it's not a pmap collective primitive, act just like batching
         rule = batching.get_primitive_batcher(primitive)
@@ -817,7 +879,7 @@ class SplitAxisTrace(core.Trace):
         else:
           return new_tracer(val_out, axis_out)
 
-  def process_call(self, call_primitive, f, tracers, params):
+  def process_call(self, call_primitive, f: lu.WrappedFun, tracers, params):
     assert call_primitive.multiple_results
     if call_primitive in pe.map_primitives:
       return self.process_map(call_primitive, f, tracers, params)
@@ -830,7 +892,7 @@ class SplitAxisTrace(core.Trace):
         vals_out = call_primitive.bind(f, *vals, **params)
         return [SplitAxisTracer(self, a, x) for a, x in zip(names_out(), vals_out)]
 
-  def process_map(self, map_primitive, f, tracers, params):
+  def process_map(self, map_primitive, f: lu.WrappedFun, tracers, params):
     vals, names = unzip2((t.val, t.axis_name) for t in tracers)
     if all(name is not_mapped for name in names):
       return map_primitive.bind(f, *vals, **params)
@@ -855,4 +917,4 @@ class SplitAxisTrace(core.Trace):
     return  val, todo
 
 
-split_axis_rules = {}
+split_axis_rules: Dict[core.Primitive, Callable] = {}
