@@ -16,25 +16,24 @@
 Control flow primitives.
 """
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
 
 import collections
 import functools
 import itertools
 import operator
 import threading
+from typing import Callable
 
 import numpy as onp
 
-from jax import api
+import jax
 from jax import core
 from jax import dtypes
 from jax.lax import lax
 from jax import linear_util as lu
 from jax.abstract_arrays import ShapedArray, raise_to_shaped
 from jax.api_util import flatten_fun_nokwargs, apply_flat_fun_nokwargs
+from jax.core import get_aval
 from jax.interpreters import ad
 from jax.interpreters import partial_eval as pe
 from jax.interpreters import xla
@@ -43,9 +42,10 @@ from jax.interpreters import masking
 from jax.lib import xla_bridge as xb
 from jax.lib import xla_client
 from jax.util import (partial, unzip2, safe_map, safe_zip, split_list,
-                      split_dict, cache)
+                      split_dict, cache, extend_name_stack)
 from jax.tree_util import (tree_flatten, tree_unflatten, treedef_is_leaf,
-                           treedef_children, treedef_tuple)
+                           treedef_children, treedef_tuple, tree_leaves,
+                           tree_multimap)
 from jax import ad_util
 
 _map = safe_map
@@ -54,14 +54,15 @@ _reduce = functools.reduce
 
 
 @cache()
-def _initial_style_jaxpr(fun, in_tree, in_avals):
+def _initial_style_jaxpr(fun: Callable, in_tree, in_avals):
   in_pvals = [pe.PartialVal((aval, core.unit)) for aval in in_avals]
-  fun, out_tree = flatten_fun_nokwargs(lu.wrap_init(fun), in_tree)
-  jaxpr, out_pvals, consts = pe.trace_to_jaxpr(fun, in_pvals, instantiate=True,
-                                               stage_out_calls=True)
+  wrapped_fun, out_tree = flatten_fun_nokwargs(lu.wrap_init(fun), in_tree)
+  with core.initial_style_staging():
+    jaxpr, out_pvals, consts = pe.trace_to_jaxpr(
+      wrapped_fun, in_pvals, instantiate=True, stage_out=False)
   out_avals = _map(raise_to_shaped, unzip2(out_pvals)[0])
   const_avals = tuple(raise_to_shaped(core.get_aval(c)) for c in consts)
-  typed_jaxpr = core.TypedJaxpr(pe.closure_convert_jaxpr(jaxpr),
+  typed_jaxpr = core.TypedJaxpr(pe.convert_constvars_jaxpr(jaxpr),
                                 (), const_avals + in_avals, out_avals)
   return typed_jaxpr, consts, out_tree()
 
@@ -79,7 +80,11 @@ def typematch(aval1, aval2):
   return (raise_to_shaped(aval1).strip_weak_type() ==
           raise_to_shaped(aval2).strip_weak_type())
 
-class FixedPointError(Exception): pass
+def _disable_jit_impl(prim, interp, *args, **kwargs):
+  if jax.api._jit_is_disabled():
+    return interp(*args, **kwargs)
+  else:
+    return xla.apply_primitive(prim, *args, **kwargs)
 
 
 ### fori_loop and while_loop
@@ -94,6 +99,13 @@ def _fori_body_fun(body_fun):
     i, upper, x = loop_carry
     return lax.add(i, lax._const(i, 1)), upper, body_fun(i, x)
   return while_body_fun
+
+@cache()
+def _fori_scan_body_fun(body_fun):
+  def scanned_fun(loop_carry, _):
+    i, upper, x = loop_carry
+    return (lax.add(i, lax._const(i, 1)), upper, body_fun(i, x)), None
+  return scanned_fun
 
 def fori_loop(lower, upper, body_fun, init_val):
   """Loop from ``lower`` to ``upper`` by reduction to ``while_loop``.
@@ -132,15 +144,31 @@ def fori_loop(lower, upper, body_fun, init_val):
   Returns:
     Loop value from the final iteration, of type ``a``.
   """
-  # TODO: perhaps do more type checking here, for better error messages.
+  # TODO(phawkins): perhaps do more type checking here, better error messages.
   lower_dtype = dtypes.canonicalize_dtype(lax.dtype(lower))
   upper_dtype = dtypes.canonicalize_dtype(lax.dtype(upper))
   if lower_dtype != upper_dtype:
     msg = ("lower and upper arguments to fori_loop must have equal types, "
            "got {} and {}")
     raise TypeError(msg.format(lower_dtype.name, upper_dtype.name))
-  _, _, result = while_loop(_fori_cond_fun, _fori_body_fun(body_fun),
-                            (lower, upper, init_val))
+
+  # If we can specialize on the trip count, call scan instead of a while_loop
+  # to enable efficient reverse-mode differentiation.
+  try:
+    lower_ = int(lower)
+    upper_ = int(upper)
+  except TypeError:
+    use_scan = False
+  else:
+    use_scan = False  # TODO(mattjj): re-enable this
+
+  if use_scan:
+    (_, _, result), _ = scan(_fori_scan_body_fun(body_fun),
+                             (lower, upper, init_val), None,
+                             length=upper_ - lower_)
+  else:
+    _, _, result = while_loop(_fori_cond_fun, _fori_body_fun(body_fun),
+                              (lower, upper, init_val))
   return result
 
 
@@ -188,6 +216,12 @@ def while_loop(cond_fun, body_fun, init_val):
   Returns:
     The output from the final iteration of body_fun, of type ``a``.
   """
+  if jax.api._jit_is_disabled():
+    val = init_val
+    while cond_fun(val):
+      val = body_fun(val)
+    return val
+
   init_vals, in_tree = tree_flatten((init_val,))
   init_avals = tuple(_map(_abstractify, init_vals))
   cond_jaxpr, cond_consts, cond_tree = _initial_style_jaxpr(cond_fun, in_tree, init_avals)
@@ -213,10 +247,8 @@ def while_loop(cond_fun, body_fun, init_val):
 def _while_loop_abstract_eval(*args, **kwargs):
   return _map(raise_to_shaped, kwargs["body_jaxpr"].out_avals)
 
-def _while_loop_translation_rule(c, axis_env, *args, **kwargs):
-  backend = kwargs.pop('backend')
-  cond_jaxpr, body_jaxpr, cond_nconsts, body_nconsts = split_dict(
-      kwargs, ["cond_jaxpr", "body_jaxpr", "cond_nconsts", "body_nconsts"])
+def _while_loop_translation_rule(c, axis_env, name_stack, avals, backend, *args,
+                                 cond_jaxpr, body_jaxpr, cond_nconsts, body_nconsts):
   cond_consts, body_consts, init_vals = split_list(args, [cond_nconsts, body_nconsts])
   batched = bool(cond_jaxpr.out_avals[0].shape)
 
@@ -233,7 +265,8 @@ def _while_loop_translation_rule(c, axis_env, *args, **kwargs):
   cond_carry_elts = [cond_c.GetTupleElement(cond_carry, i) for i in range(len(args))]
   x, _, z = split_list(cond_carry_elts, [cond_nconsts, body_nconsts])
   pred, = xla.jaxpr_subcomp(cond_c, cond_jaxpr.jaxpr, backend, axis_env,
-                            _map(cond_c.Constant, cond_jaxpr.literals), (), *(x + z))
+                            _map(cond_c.Constant, cond_jaxpr.literals),
+                            extend_name_stack(name_stack, 'cond'), *(x + z))
   if batched:
     scalar = ShapedArray((), onp.bool_)
     or_ = xla.primitive_subcomputation(lax.or_p, scalar, scalar)
@@ -245,10 +278,12 @@ def _while_loop_translation_rule(c, axis_env, *args, **kwargs):
   body_carry_elts = [body_c.GetTupleElement(body_carry, i) for i in range(len(args))]
   x, y, z = split_list(body_carry_elts, [cond_nconsts, body_nconsts])
   new_z = xla.jaxpr_subcomp(body_c, body_jaxpr.jaxpr, backend, axis_env,
-                            _map(body_c.Constant, body_jaxpr.literals), (), *(y + z))
+                            _map(body_c.Constant, body_jaxpr.literals),
+                            extend_name_stack(name_stack, 'body'), *(y + z))
   if batched:
     body_pred, = xla.jaxpr_subcomp(body_c, cond_jaxpr.jaxpr, backend, axis_env,
-                                   _map(body_c.Constant, cond_jaxpr.literals), (), *(x + z))
+                                   _map(body_c.Constant, cond_jaxpr.literals),
+                                   extend_name_stack(name_stack, 'body_pred'), *(x + z))
     new_z = _map(partial(_pred_bcast_select, body_c, body_pred), new_z, z)
     assert _map(body_c.GetShape, new_z) == _map(body_c.GetShape, z) # no broadcast
   new_carry = body_c.Tuple(*itertools.chain(x, y, new_z))
@@ -341,9 +376,8 @@ def _while_loop_jvp(primals, tangents, cond_nconsts, cond_jaxpr, body_nconsts,
 
   newvar = core.gensym('')
   invars_aug = (
-      cond_jaxpr.jaxpr.invars + [newvar() for _ in range(len(init_dot))])
+      cond_jaxpr.jaxpr.invars + [newvar(get_aval(x)) for x in init_dot])
   cond_jaxpr_augmented = core.Jaxpr(cond_jaxpr.jaxpr.constvars,
-                                    cond_jaxpr.jaxpr.freevars,
                                     invars_aug,
                                     cond_jaxpr.jaxpr.outvars,
                                     cond_jaxpr.jaxpr.eqns)
@@ -398,9 +432,9 @@ def cond(pred, true_operand, true_fun, false_operand, false_fun):
 
   try:
     pred_dtype = dtypes.result_type(pred)
-  except TypeError:
+  except TypeError as err:
     msg = ("Pred type must be either boolean or number, got {}.")
-    raise TypeError(msg.format(pred))
+    raise TypeError(msg.format(pred)) from err
 
   if pred_dtype.kind != 'b':
     if pred_dtype.kind in 'iuf':
@@ -408,6 +442,13 @@ def cond(pred, true_operand, true_fun, false_operand, false_fun):
     else:
       msg = ("Pred type must be either boolean or number, got {}.")
       raise TypeError(msg.format(pred_dtype))
+
+  if jax.api._jit_is_disabled():
+    if pred:
+      return true_fun(true_operand)
+    else:
+      return false_fun(false_operand)
+
   true_ops, true_tree = tree_flatten((true_operand,))
   true_avals = tuple(_map(_abstractify, true_ops))
   true_jaxpr, true_consts, true_out_tree = _initial_style_jaxpr(true_fun, true_tree, true_avals)
@@ -417,66 +458,65 @@ def cond(pred, true_operand, true_fun, false_operand, false_fun):
   _check_tree_and_avals("true_fun and false_fun output",
                         true_out_tree, true_jaxpr.out_avals,
                         false_out_tree, false_jaxpr.out_avals)
+  linear = (False,) * (len(true_consts) + len(true_ops) + len(false_consts) +
+                       len(false_ops))
   out = cond_p.bind(
       *itertools.chain([pred], true_consts, true_ops, false_consts, false_ops),
-      true_jaxpr=true_jaxpr, false_jaxpr=false_jaxpr,
-      true_nconsts=len(true_consts), false_nconsts=len(false_consts))
+      true_jaxpr=true_jaxpr, false_jaxpr=false_jaxpr, linear=linear)
   return tree_unflatten(true_out_tree, out)
 
 def _cond_abstract_eval(*args, **kwargs):
   return _map(raise_to_shaped, kwargs["true_jaxpr"].out_avals)
 
-def _cond_translation_rule(c, axis_env, pred, *args, **kwargs):
-  backend = kwargs.pop("backend", None)
-  true_jaxpr, false_jaxpr, true_nconsts, false_nconsts = split_dict(
-      kwargs, ["true_jaxpr", "false_jaxpr", "true_nconsts", "false_nconsts"])
-  true_nops = len(true_jaxpr.in_avals) - true_nconsts
-  true_consts, true_ops, false_consts, false_ops = split_list(
-      args, [true_nconsts, true_nops, false_nconsts])
+def _cond_translation_rule(c, axis_env, name_stack, avals, backend,
+                           pred, *args, true_jaxpr, false_jaxpr, linear):
+  del linear  # Unused.
+  true_ops, false_ops = split_list(args, [len(true_jaxpr.in_avals)])
 
   def make_computation(name, jaxpr, op_shape):
-    c = xb.make_computation_builder(name)
+    c = xb.make_computation_builder(name + '_comp')
     op = c.ParameterWithShape(op_shape)
     ops = [c.GetTupleElement(op, i) for i in range(len(jaxpr.in_avals))]
     outs = xla.jaxpr_subcomp(c, jaxpr.jaxpr, backend, axis_env,
-                             _map(c.Constant, jaxpr.literals), (), *ops)
+                             _map(c.Constant, jaxpr.literals),
+                             extend_name_stack(name_stack, name + '_fun'), *ops)
     return c.Build(c.Tuple(*outs))
 
-  true_op = c.Tuple(*(true_consts + true_ops))
-  true_c = make_computation("true_comp", true_jaxpr, c.GetShape(true_op))
+  true_op = c.Tuple(*true_ops)
+  true_c = make_computation('true', true_jaxpr, c.GetShape(true_op))
 
-  false_op = c.Tuple(*(false_consts + false_ops))
-  false_c = make_computation("false_comp", false_jaxpr, c.GetShape(false_op))
+  false_op = c.Tuple(*false_ops)
+  false_c = make_computation('false', false_jaxpr, c.GetShape(false_op))
 
   return c.Conditional(pred, true_op, true_c, false_op, false_c)
 
 def _cond_pred_bcast_select(pred, x, y):
-  bcast_pred = lax.broadcast_in_dim(pred, onp.shape(x), list(range(onp.ndim(pred))))
-  return lax.select(bcast_pred, x, y)
+  if core.get_aval(x) is core.get_aval(y) is core.abstract_unit:
+    return x
+  else:
+    bcast_pred = lax.broadcast_in_dim(pred, onp.shape(x), list(range(onp.ndim(pred))))
+    return lax.select(bcast_pred, x, y)
 
-def _cond_batching_rule(args, dims, true_jaxpr, false_jaxpr, true_nconsts,
-                        false_nconsts):
+def _cond_batching_rule(args, dims, true_jaxpr, false_jaxpr, linear):
   # TODO: maybe avoid moving arg axes to front if we're promoting to select?
+  size, = {x.shape[d] for x, d in zip(args, dims) if d is not batching.not_mapped}
   args = [batching.moveaxis(x, d, 0) if d is not batching.not_mapped and d != 0
           else x for x, d in zip(args, dims)]
-  true_nops = len(true_jaxpr.in_avals) - true_nconsts
-  (pred,), true_consts, true_ops, false_consts, false_ops = split_list(
-      args, [1, true_nconsts, true_nops, false_nconsts])
-  size, = {x.shape[d] for x, d in zip(args, dims) if d is not batching.not_mapped}
   orig_bat = [d is not batching.not_mapped for d in dims]
-  (pred_bat,), tconst_bat, t_bat, fconst_bat, f_bat = split_list(
-    orig_bat, [1, true_nconsts, true_nops, false_nconsts])
+  del dims
+  (pred,), true_ops, false_ops = split_list(args, [1, len(true_jaxpr.in_avals)])
+  (pred_bat,), t_bat, f_bat = split_list(orig_bat, [1, len(true_jaxpr.in_avals)])
 
-  _, true_out_bat = batching.batch_jaxpr(true_jaxpr, size, tconst_bat + t_bat, False)
-  _, false_out_bat = batching.batch_jaxpr(false_jaxpr, size, fconst_bat + f_bat, False)
+  _, true_out_bat = batching.batch_jaxpr(true_jaxpr, size, t_bat, False)
+  _, false_out_bat = batching.batch_jaxpr(false_jaxpr, size, f_bat, False)
   out_bat = [a or b for a, b in zip(true_out_bat, false_out_bat)]
 
-  true_jaxpr_batched, _ = batching.batch_jaxpr(true_jaxpr, size, tconst_bat + t_bat, out_bat)
-  false_jaxpr_batched, _ = batching.batch_jaxpr(false_jaxpr, size, fconst_bat + f_bat, out_bat)
+  true_jaxpr_batched, _ = batching.batch_jaxpr(true_jaxpr, size, t_bat, out_bat)
+  false_jaxpr_batched, _ = batching.batch_jaxpr(false_jaxpr, size, f_bat, out_bat)
 
   if pred_bat:
-    true_out = core.jaxpr_as_fun(true_jaxpr_batched)(*(true_consts + true_ops))
-    false_out = core.jaxpr_as_fun(false_jaxpr_batched)(*(false_consts + false_ops))
+    true_out = core.jaxpr_as_fun(true_jaxpr_batched)(*true_ops)
+    false_out = core.jaxpr_as_fun(false_jaxpr_batched)(*false_ops)
     true_out = [batching.broadcast(x, size, 0) if not b else x
                 for x, b in zip(true_out, out_bat)]
     false_out = [batching.broadcast(x, size, 0) if not b else x
@@ -485,68 +525,210 @@ def _cond_batching_rule(args, dims, true_jaxpr, false_jaxpr, true_nconsts,
             for t, f in zip(true_out, false_out)], [0] * len(true_out)
   else:
     out_dims = [0 if b else batching.not_mapped for b in out_bat]
-    return cond_p.bind(
-      *itertools.chain([pred], true_consts, true_ops, false_consts, false_ops),
-      true_jaxpr=true_jaxpr_batched, false_jaxpr=false_jaxpr_batched,
-      true_nconsts=len(true_consts), false_nconsts=len(false_consts)), out_dims
+    out = cond_p.bind(
+      *itertools.chain([pred], true_ops, false_ops),
+      true_jaxpr=true_jaxpr_batched, false_jaxpr=false_jaxpr_batched, linear=linear)
+    return out, out_dims
 
-def _cond_jvp(primals, tangents, true_jaxpr, false_jaxpr, true_nconsts,
-              false_nconsts):
+def _cond_jvp(primals, tangents, true_jaxpr, false_jaxpr, linear):
   nonzeros = [t is not ad_util.zero for t in tangents]
 
-  true_nops = len(true_jaxpr.in_avals) - true_nconsts
-  false_nops = len(false_jaxpr.in_avals) - false_nconsts
-
-  (pred_nz,), tconst_nz, t_nz, fconst_nz, f_nz = split_list(
-      nonzeros, [1, true_nconsts, true_nops, false_nconsts])
-
+  (pred_nz,), t_nz, f_nz = split_list(nonzeros, [1, len(true_jaxpr.in_avals)])
   assert pred_nz is False
 
-  _, true_out_nz = ad.jvp_jaxpr(true_jaxpr, tconst_nz + t_nz,
-                                instantiate=False)
-  _, false_out_nz = ad.jvp_jaxpr(false_jaxpr, fconst_nz + f_nz,
-                                 instantiate=False)
-
+  _, true_out_nz = ad.jvp_jaxpr(true_jaxpr, t_nz, instantiate=False)
+  _, false_out_nz = ad.jvp_jaxpr(false_jaxpr, f_nz, instantiate=False)
   out_nz = [a or b for a, b in zip(true_out_nz, false_out_nz)]
 
-  true_jvp, _ = ad.jvp_jaxpr(true_jaxpr, tconst_nz + t_nz, instantiate=out_nz)
-  false_jvp, _ = ad.jvp_jaxpr(false_jaxpr, fconst_nz + f_nz, instantiate=out_nz)
+  true_jvp, _ = ad.jvp_jaxpr(true_jaxpr, t_nz, instantiate=out_nz)
+  false_jvp, _ = ad.jvp_jaxpr(false_jaxpr, f_nz, instantiate=out_nz)
 
-  (pred,), tconsts, tops, fconsts, fops = split_list(
-      primals, [1, true_nconsts, true_nops, false_nconsts])
-  _, tconsts_dot, tops_dot, fconsts_dot, fops_dot = split_list(
-      tangents, [1, true_nconsts, true_nops, false_nconsts])
+  (pred,), tops, fops = split_list(primals, [1, len(true_jaxpr.in_avals)])
+  _, tops_dot, fops_dot = split_list(tangents, [1, len(true_jaxpr.in_avals)])
 
-  tconsts_dot = _prune_zeros(tconsts_dot)
   tops_dot = _prune_zeros(tops_dot)
-  fconsts_dot = _prune_zeros(fconsts_dot)
   fops_dot = _prune_zeros(fops_dot)
 
-  true_jvp = ad.rearrange_binders(true_jvp, [true_nconsts, true_nops],
-                                  [len(tconsts_dot), len(tops_dot)],
-                                  [len(out_nz)], [sum(out_nz)])
-  false_jvp = ad.rearrange_binders(false_jvp, [false_nconsts, false_nops],
-                                   [len(fconsts_dot), len(fops_dot)],
-                                   [len(out_nz)], [sum(out_nz)])
-
+  tops_lin, fops_lin = _map(tuple, split_list(linear, [len(tops)]))
+  linear_jvp = (tops_lin + (True,) * len(tops_dot) +
+                fops_lin + (True,) * len(fops_dot))
   out = cond_p.bind(
-      *itertools.chain([pred],
-                       tconsts, tconsts_dot, tops, tops_dot,
-                       fconsts, fconsts_dot, fops, fops_dot),
-      true_jaxpr=true_jvp, false_jaxpr=false_jvp,
-      true_nconsts=len(tconsts) + len(tconsts_dot),
-      false_nconsts=len(fconsts) + len(fconsts_dot))
+      *itertools.chain([pred], tops, tops_dot, fops, fops_dot),
+      true_jaxpr=true_jvp, false_jaxpr=false_jvp, linear=linear_jvp)
   out_primals, out_tangents = split_list(out, [len(out_nz)])
   out_tangents_iter = iter(out_tangents)
   out_tangents = [
       next(out_tangents_iter) if nz else ad_util.zero for nz in out_nz]
   return out_primals, out_tangents
 
+def _cond_partial_eval(trace, *tracers, true_jaxpr, false_jaxpr, linear):
+  unknowns = [t.pval[0] is not None for t in tracers]
+
+  (pred_uk,), t_uk, f_uk = split_list(unknowns, [1, len(true_jaxpr.in_avals)])
+
+  if pred_uk:
+    # When the predicate is unknown, we stage out the whole cond.
+    params = dict(true_jaxpr=true_jaxpr, false_jaxpr=false_jaxpr, linear=linear)
+    return trace.default_process_primitive(cond_p, tracers, params)
+
+  _, _, t_out_uks = pe.partial_eval_jaxpr(true_jaxpr, t_uk, instantiate=False)
+  _, _, f_out_uks = pe.partial_eval_jaxpr(false_jaxpr, f_uk, instantiate=False)
+  out_uks = [a or b for a, b in zip(t_out_uks, f_out_uks)]
+
+  true_jaxpr_1, true_jaxpr_2, _ = pe.partial_eval_jaxpr(true_jaxpr, t_uk,
+                                                        instantiate=out_uks)
+  false_jaxpr_1, false_jaxpr_2, _ = pe.partial_eval_jaxpr(false_jaxpr, f_uk,
+                                                          instantiate=out_uks)
+
+  num_t_res = len(true_jaxpr_1.out_avals) - len(out_uks)
+  num_f_res = len(false_jaxpr_1.out_avals) - len(out_uks)
+
+  move = [False] * len(true_jaxpr.in_avals) + [True] * num_t_res
+  true_jaxpr_2 = pe.move_binders_to_front(true_jaxpr_2, move)
+  move = [False] * len(false_jaxpr.in_avals) + [True] * num_f_res
+  false_jaxpr_2 = pe.move_binders_to_front(false_jaxpr_2, move)
+
+  # TODO(frostig,mattjj): pe.partial_eval_jaxpr should raise to shaped avals
+  t_res_avals = _map(raise_to_shaped, true_jaxpr_2.in_avals[:num_t_res])
+  f_res_avals = _map(raise_to_shaped, false_jaxpr_2.in_avals[:num_f_res])
+
+  assert len(true_jaxpr_2.out_avals) == len(false_jaxpr_2.out_avals)
+  num_outs = len(true_jaxpr_2.out_avals)
+
+  true_jaxpr_1 = _join_cond_outputs(
+      true_jaxpr_1, num_outs, f_res_avals, zeros_on_left=False)
+  false_jaxpr_1 = _join_cond_outputs(
+      false_jaxpr_1, num_outs, t_res_avals, zeros_on_left=True)
+
+  # TODO(frostig,mattjj): reinstate this assertion once pe.partial_eval_jaxpr
+  # raises to shaped avals
+  # assert true_jaxpr_1.out_avals == false_jaxpr_1.out_avals
+  num_res = num_t_res + num_f_res
+
+  _, in_consts = unzip2([t.pval for t in tracers])
+  out_consts_res = cond_p.bind(
+      *in_consts, true_jaxpr=true_jaxpr_1, false_jaxpr=false_jaxpr_1,
+      linear=linear)
+  out_consts, res = split_list(out_consts_res, [len(out_consts_res) - num_res])
+
+  # TODO(frostig,mattjj): remove raised_to_shaped of avals once
+  # pe.partial_eval_jaxpr handles it
+  out_avals = _map(raise_to_shaped, true_jaxpr_2.out_avals)
+  out_pvs = [aval if uk else None for aval, uk in zip(out_avals, out_uks)]
+
+  pred_tracer = trace.instantiate_const(tracers[0])
+
+  ops_tracers = [trace.instantiate_const(t) if uk
+                 else trace.new_instantiated_literal(core.unit)
+                 for uk, t in zip(unknowns[1:], tracers[1:])]
+  true_ops_tracers, false_ops_tracers = split_list(
+      ops_tracers, [len(true_jaxpr.in_avals)])
+
+  res_tracers = _map(trace.new_instantiated_const, res)
+  true_res_tracers, false_res_tracers = split_list(res_tracers, [num_t_res])
+
+  out_tracers = [pe.JaxprTracer(trace, pe.PartialVal((pv, const)), None)
+                 for pv, const in zip(out_pvs, out_consts)]
+
+  tops_lin, fops_lin = _map(tuple, split_list(linear, [len(true_jaxpr.in_avals)]))
+  linear_2 = ((False,) * num_t_res + tops_lin + (False,) * num_f_res + fops_lin)
+  params = dict(true_jaxpr=true_jaxpr_2, false_jaxpr=false_jaxpr_2,
+                linear=linear_2)
+  eqn = pe.new_eqn_recipe([pred_tracer] +
+                          true_res_tracers + true_ops_tracers +
+                          false_res_tracers + false_ops_tracers,
+                          out_tracers,
+                          cond_p, params)
+  for t in out_tracers: t.recipe = eqn
+  return out_tracers
+
+def _join_cond_outputs(jaxpr, num_prefix, zeros_avals, zeros_on_left):
+  @lu.wrap_init
+  def f_aug(*args):
+    prefix_and_rest = core.jaxpr_as_fun(jaxpr)(*args)
+    prefix, rest = split_list(prefix_and_rest, [num_prefix])
+    zeros = [ad_util.zeros_like_aval(a) for a in zeros_avals]
+    if zeros_on_left:
+      return prefix + zeros + rest
+    else:
+      return prefix + rest + zeros
+
+  return _make_typed_jaxpr(f_aug, jaxpr.in_avals)
+
+def _transpose_cond_jaxpr(jaxpr, num_res):
+  num_non_res = len(jaxpr.in_avals) - num_res
+  res_avals, primal_avals = split_list(jaxpr.in_avals, [num_res])
+  primal_avals = _map(raise_to_shaped, primal_avals)
+
+  @lu.wrap_init
+  def transposed(*args):
+    res, cts_out = split_list(args, [num_res])
+    primals = res + [ad.UndefinedPrimal(aval) for aval in primal_avals]
+    cts_in = ad.backward_pass(
+        jaxpr.jaxpr, jaxpr.literals, primals, cts_out)
+    _, cts_in = split_list(cts_in, [num_res])
+    return _map(ad.instantiate_zeros_aval, primal_avals, cts_in)
+
+  return _make_typed_jaxpr(transposed, res_avals + jaxpr.out_avals)
+
+def _cond_transpose(cts, *args, true_jaxpr, false_jaxpr, linear):
+  (pred,), tops, fops = split_list(args, [1, len(true_jaxpr.in_avals)])
+  tops_lin, fops_lin = split_list(linear, [len(true_jaxpr.in_avals)])
+  in_avals = _map(raise_to_shaped, true_jaxpr.in_avals + false_jaxpr.in_avals)
+
+  num_t_res = len(tops) - sum(tops_lin)
+  num_f_res = len(fops) - sum(fops_lin)
+
+  t_jaxpr_trans = _transpose_cond_jaxpr(true_jaxpr, num_t_res)
+  f_jaxpr_trans = _transpose_cond_jaxpr(false_jaxpr, num_f_res)
+  lin_in_avals = _map(raise_to_shaped, [a for a, l in zip(in_avals, linear) if l])
+  assert t_jaxpr_trans.out_avals + f_jaxpr_trans.out_avals == lin_in_avals
+
+  t_jaxpr_trans_ = _join_cond_outputs(
+      t_jaxpr_trans, 0, f_jaxpr_trans.out_avals, zeros_on_left=False)
+  f_jaxpr_trans_ = _join_cond_outputs(
+      f_jaxpr_trans, 0, t_jaxpr_trans.out_avals, zeros_on_left=True)
+  assert t_jaxpr_trans_.out_avals == f_jaxpr_trans_.out_avals == lin_in_avals
+
+  t_res, _ = split_list(tops, [num_t_res])
+  f_res, _ = split_list(fops, [num_f_res])
+
+  linear_trans = ((False,) * num_t_res + (True,) * len(cts) +
+                  (False,) * num_f_res + (True,) * len(cts))
+
+  cts = _map(ad.instantiate_zeros_aval, true_jaxpr.out_avals, cts)
+
+  out = cond_p.bind(
+      pred, *itertools.chain(t_res, cts, f_res, cts),
+      true_jaxpr=t_jaxpr_trans_, false_jaxpr=f_jaxpr_trans_,
+      linear=linear_trans)
+  assert all(_map(typecheck, lin_in_avals, out))
+
+  out_iter = iter(out)
+  out = [next(out_iter) if l else None for l in linear]
+  assert next(out_iter, None) is None
+  return [None] + out
+
+def cond_bind(*args, true_jaxpr, false_jaxpr, linear):
+  if not core.skip_checks:
+    assert len(linear) + 1 == len(args)
+    assert len(args) == 1 + len(true_jaxpr.in_avals) + len(false_jaxpr.in_avals)
+    (pred,), tops, fops = split_list(args, [1, len(true_jaxpr.in_avals)])
+    assert all(_map(typecheck, true_jaxpr.in_avals, tops))
+    assert all(_map(typecheck, false_jaxpr.in_avals, fops))
+    core.check_jaxpr(true_jaxpr.jaxpr)
+    core.check_jaxpr(false_jaxpr.jaxpr)
+  return core.Primitive.bind(cond_p, *args, true_jaxpr=true_jaxpr,
+                             false_jaxpr=false_jaxpr, linear=linear)
+
 cond_p = lax.Primitive('cond')
 cond_p.multiple_results = True
 cond_p.def_impl(partial(xla.apply_primitive, cond_p))
 cond_p.def_abstract_eval(_cond_abstract_eval)
+cond_p.def_custom_bind(cond_bind)
 ad.primitive_jvps[cond_p] = _cond_jvp
+ad.primitive_transposes[cond_p] = _cond_transpose
+pe.custom_partial_eval_rules[cond_p] = _cond_partial_eval
 batching.primitive_batchers[cond_p] = _cond_batching_rule
 xla.initial_style_translations[cond_p] = _cond_translation_rule
 
@@ -620,15 +802,16 @@ def scan(f, init, xs, length=None):
     the second output of ``f`` when scanned over the leading axis of the inputs.
   """
   init_flat, init_tree = tree_flatten(init)
-  xs_flat, _ = tree_flatten(xs)
+  xs_flat, xs_tree = tree_flatten(xs)
   in_flat, in_tree = tree_flatten((init, xs))
 
   try:
     lengths = [x.shape[0] for x in xs_flat]
-  except AttributeError:
+  except AttributeError as err:
     msg = "scan got value with no leading axis to scan over: {}."
-    raise ValueError(msg.format(', '.join(str(x) for x in xs_flat
-                                          if not hasattr(x, 'shape'))))
+    raise ValueError(
+      msg.format(', '.join(str(x) for x in xs_flat
+                           if not hasattr(x, 'shape')))) from err
 
   if length is not None:
     length = int(length)
@@ -646,6 +829,17 @@ def scan(f, init, xs, length=None):
       raise ValueError(msg)
     else:
       length, = unique_lengths
+
+  if jax.api._jit_is_disabled():
+    carry = init
+    ys = []
+    for i in range(length):
+      xs_slice = [_index_array(i, core.get_aval(x), x) for x in xs_flat]
+      carry, y = f(carry, tree_unflatten(xs_tree, xs_slice))
+      ys.append(y)
+    stack = lambda y, *ys: (y if core.get_aval(y) is core.abstract_unit
+                            else jax.numpy.stack((y, *ys)))
+    return carry, tree_multimap(stack, *ys)
 
   carry_avals = tuple(_map(_abstractify, init_flat))
   x_shapes = [masking.padded_shape_as_value(x.shape[1:]) for x in xs_flat]
@@ -667,25 +861,31 @@ def scan(f, init, xs, length=None):
                     linear=(False,) * (len(consts) + len(in_flat)))
   return tree_unflatten(out_tree, out)
 
-def _scan_impl(*args, **kwargs):
-  forward, length, num_consts, num_carry, jaxpr, linear = split_dict(
-      kwargs, ["forward", "length", "num_consts", "num_carry", "jaxpr", "linear"])
-
+def _scan_impl(*args, forward, length, num_consts, num_carry, jaxpr, linear):
   consts, init, xs = split_list(args, [num_consts, num_carry])
   _, _, x_avals = split_list(jaxpr.in_avals, [num_consts, num_carry])
   _, y_avals = split_list(jaxpr.out_avals, [num_carry])
 
-  def body_fun(i, vals):
-    i = i if forward else length - i - 1
-    carry, ys = split_list(vals, [num_carry])
-    x = _map(partial(_index_array, i), x_avals, xs)
+  def cond_fun(vals):
+    i, *_ = vals
+    return i < length
+
+  def body_fun(vals):
+    [i], carry, ys = split_list(vals, [1, num_carry])
+    i_ = i if forward else length - i - 1
+    x = _map(partial(_index_array, i_), x_avals, xs)
     out_flat = core.jaxpr_as_fun(jaxpr)(*(consts + carry + x))
     carry_out, y_updates = split_list(out_flat, [num_carry])
-    ys_out = _map(partial(_update_array, i), y_avals, ys, y_updates)
-    return carry_out + ys_out
+    ys_out = _map(partial(_update_array, i_), y_avals, ys, y_updates)
+    return [i + 1] + carry_out + ys_out
 
   ys_init = _map(partial(_empty_array, length), y_avals)
-  return fori_loop(lax._const(length, 0), length, body_fun, init + ys_init)
+  if length == 0:
+    return init + ys_init
+  else:
+    init_val = [lax._const(length, 0)] + init + ys_init
+    _, *outs = while_loop(cond_fun, body_fun, init_val)
+    return outs
 
 def _index_array(i, aval, x):
   if aval is core.abstract_unit:
@@ -704,6 +904,12 @@ def _update_array(i, aval, xs, x):
     return core.unit
   else:
     return lax.dynamic_update_index_in_dim(xs, x, i, 0)
+
+def _scan_abstract_eval(*args, forward, length, num_consts, num_carry, jaxpr, linear):
+  carry_avals, y_avals = split_list(jaxpr.out_avals, [num_carry])
+  ys_avals = [ShapedArray((length,) + aval.shape, aval.dtype)
+              if aval is not core.abstract_unit else aval for aval in y_avals]
+  return carry_avals + ys_avals
 
 def _scan_jvp(primals, tangents, forward, length, jaxpr, num_consts, num_carry,
               linear):
@@ -743,9 +949,9 @@ def _scan_jvp(primals, tangents, forward, length, jaxpr, num_consts, num_carry,
       [num_carry, num_ys], [len(init_dot), sum(nonzeros_out) - len(init_dot)])
 
   consts_linear, init_linear, xs_linear = split_list(linear, [num_consts, num_carry])
-  jaxpr_jvp_linear = (consts_linear + [True] * len(consts_dot)
-                      + init_linear + [True] * len(init_dot)
-                      + xs_linear + [True] * len(xs_dot))
+  jaxpr_jvp_linear = tuple(consts_linear + [True] * len(consts_dot)
+                           + init_linear + [True] * len(init_dot)
+                           + xs_linear + [True] * len(xs_dot))
 
   out_flat = scan_p.bind(
       *(consts + consts_dot + init + init_dot + xs + xs_dot),
@@ -763,9 +969,8 @@ def _scan_jvp(primals, tangents, forward, length, jaxpr, num_consts, num_carry,
 def _prune_zeros(ts):
   return [t for t in ts if t is not ad_util.zero]
 
-def _scan_partial_eval(trace, *tracers, **kwargs):
-  forward, length, num_consts, num_carry, jaxpr, linear = split_dict(
-      kwargs, ["forward", "length", "num_consts", "num_carry", "jaxpr", "linear"])
+def _scan_partial_eval(trace, *tracers, forward, length, num_consts, num_carry,
+                       jaxpr, linear):
   num_xs = len(jaxpr.in_avals) - num_carry - num_consts
   num_ys = len(jaxpr.out_avals) - num_carry
 
@@ -804,7 +1009,12 @@ def _scan_partial_eval(trace, *tracers, **kwargs):
   const_avals_1 = [raise_to_shaped(core.get_aval(c)) for c in consts_1]
   in_avals_1 = [core.abstract_unit] * num_consts + jaxpr_1.in_avals[num_consts:]
   out_avals_1 = [core.abstract_unit if pv is None else pv for pv, c in out_pvals_1]
-  jaxpr_1_opt = pe.TypedJaxpr(pe.closure_convert_jaxpr(untyped_jaxpr_1),
+
+  # TODO(cjfj): Explain the need for the code below.
+  for var in untyped_jaxpr_1.invars[:num_consts]:
+    var.aval = core.abstract_unit
+
+  jaxpr_1_opt = pe.TypedJaxpr(pe.convert_constvars_jaxpr(untyped_jaxpr_1),
                               (), const_avals_1 + in_avals_1, out_avals_1)
   num_consts_1 = num_consts + len(consts_1)
   # any now-known residuals are intensive, so we want to revise jaxpr_2 to take
@@ -819,10 +1029,11 @@ def _scan_partial_eval(trace, *tracers, **kwargs):
                [core.unit if uk else t.pval[1]
                 for uk, t in zip(unknowns[num_consts:], tracers[num_consts:])])
   linear_1 = ([False] * len(consts_1) + [True] * num_consts +
-              [lin or uk for uk, lin in zip(unknowns[num_consts:], linear[num_consts:])])
+              [lin or uk for uk, lin
+               in zip(unknowns[num_consts:], linear[num_consts:])])
   out_flat = scan_p.bind(
       *in_consts, forward=forward, length=length, jaxpr=jaxpr_1_opt,
-      num_consts=num_consts_1, num_carry=num_carry, linear=linear_1)
+      num_consts=num_consts_1, num_carry=num_carry, linear=tuple(linear_1))
   out_carry, ys, res_and_units = split_list(out_flat, [num_carry, num_ys])
   extensive_residuals = [r for r, (pv, _) in zip(res_and_units, res_pvals) if pv is not None]
 
@@ -842,10 +1053,10 @@ def _scan_partial_eval(trace, *tracers, **kwargs):
               [lin or not uk for uk, lin in zip(unknowns, linear)] +
               [False] * len(ext_res_tracers))
   eqn = pe.new_eqn_recipe(int_res_tracers + new_tracers + ext_res_tracers,
-                          out_tracers, scan_p, (),
+                          out_tracers, scan_p,
                           dict(forward=forward, length=length, jaxpr=jaxpr_2_opt,
                                num_consts=num_consts_2,
-                               num_carry=num_carry, linear=linear_2))
+                               num_carry=num_carry, linear=tuple(linear_2)))
   for t in out_tracers: t.recipe = eqn
   return out_tracers
 
@@ -855,10 +1066,7 @@ def _promote_aval_rank(sz, aval):
   else:
     return ShapedArray((sz,) + aval.shape, aval.dtype)
 
-def _scan_transpose(cts, *args, **kwargs):
-  forward, length, num_consts, num_carry, jaxpr, linear = split_dict(
-      kwargs, ["forward", "length", "num_consts", "num_carry", "jaxpr", "linear"])
-
+def _scan_transpose(cts, *args, forward, length, num_consts, num_carry, jaxpr, linear):
   # we've only implemented transposing scans with specific lin/nonlin patterns
   consts_lin, init_lin, xs_lin = split_list(linear, [num_consts, num_carry])
   num_ires = len(consts_lin) - sum(consts_lin)
@@ -868,13 +1076,13 @@ def _scan_transpose(cts, *args, **kwargs):
   if xs_lin != [True] * (len(xs_lin) - num_eres) + [False] * num_eres:
     raise NotImplementedError
   if not all(init_lin):
-    raise NotImplementedError
+    pass  # TODO(mattjj): error check https://github.com/google/jax/issues/1963
 
-  consts, init, xs = split_list(args, [num_consts, num_carry])
-  ires, consts = split_list(consts, [num_ires])
-  xs, eres = split_list(xs, [sum(xs_lin)])
-  assert not any(r is ad.undefined_primal for r in ires)
-  assert not any(r is ad.undefined_primal for r in eres)
+  consts, _, xs = split_list(args, [num_consts, num_carry])
+  ires, _ = split_list(consts, [num_ires])
+  _, eres = split_list(xs, [sum(xs_lin)])
+  assert not any(ad.is_undefined_primal(r) for r in ires)
+  assert not any(ad.is_undefined_primal(r) for r in eres)
 
   carry_avals, y_avals = split_list(jaxpr.out_avals, [num_carry])
   ys_avals = _map(partial(_promote_aval_rank, length), y_avals)
@@ -885,7 +1093,8 @@ def _scan_transpose(cts, *args, **kwargs):
 
   #       jaxpr :: [ires, T d] -> [T c] -> [T a, eres] -> ([T c], [T b])
   # jaxpr_trans :: [ires] -> [CT d, CT c] -> [CT b, eres] -> ([CT d, CT c], [CT a])
-  jaxpr_trans = _transpose_jaxpr(num_ires, num_consts - num_ires, num_eres, jaxpr)
+  jaxpr_trans = _transpose_scan_jaxpr(
+      num_ires, num_consts - num_ires, num_eres, jaxpr)
   linear_trans = ([False] * num_ires +
                   [True] * (len(ct_consts) + len(ct_carry) + len(ct_ys)) +
                   [False] * num_eres)
@@ -893,13 +1102,13 @@ def _scan_transpose(cts, *args, **kwargs):
   outs = scan_p.bind(
       *(ires + ct_consts + ct_carry + ct_ys + eres), forward=not forward,
       length=length, jaxpr=jaxpr_trans, num_consts=num_ires,
-      num_carry=num_consts-num_ires+num_carry, linear=linear_trans)
+      num_carry=num_consts-num_ires+num_carry, linear=tuple(linear_trans))
   ct_consts, ct_init, ct_xs = split_list(outs, [num_consts - num_ires, num_carry])
   return [None] * num_ires + ct_consts + ct_init + ct_xs + [None] * num_eres
 
-# transpose_jaxpr :: ([res1, c, a, res2] -> b)
-#                    -> ([res1, CT c, CT b, res2] -> [CT c, CT a])
-def _transpose_jaxpr(num_res1, num_c, num_res2, jaxpr):
+# transpose_scan_jaxpr :: ([res1, c, a, res2] -> b)
+#                         -> ([res1, CT c, CT b, res2] -> [CT c, CT a])
+def _transpose_scan_jaxpr(num_res1, num_c, num_res2, jaxpr):
   num_a = len(jaxpr.in_avals) - num_res1 - num_c - num_res2
   res1_avals, c_avals, a_avals, res2_avals = split_list(
       jaxpr.in_avals, [num_res1, num_c, num_a])
@@ -910,8 +1119,9 @@ def _transpose_jaxpr(num_res1, num_c, num_res2, jaxpr):
   def transposed(*res1_cbar_bbar_res2):
     res1, c_bar, b_bar, res2 = split_list(
         res1_cbar_bbar_res2, [num_res1, num_c, num_b])
-    primals = res1 + [ad.undefined_primal] * (num_c + num_a) + res2
-    _, cbar_abar = ad.backward_pass(jaxpr.jaxpr, jaxpr.literals, (), primals,
+    primals = (res1 + [ad.UndefinedPrimal(aval) for aval in c_avals] +
+               [ad.UndefinedPrimal(aval) for aval in a_avals] + res2)
+    cbar_abar = ad.backward_pass(jaxpr.jaxpr, jaxpr.literals, primals,
                                     b_bar)
     _, new_c_bar, a_bar, _ = split_list(cbar_abar, [num_res1, num_c, num_a])
     a_bar = _map(ad.instantiate_zeros_aval, a_avals, a_bar)
@@ -920,11 +1130,11 @@ def _transpose_jaxpr(num_res1, num_c, num_res2, jaxpr):
     return c_bar + a_bar
   return _make_typed_jaxpr(transposed, res1_avals + c_avals + b_avals + res2_avals)
 
-def _make_typed_jaxpr(traceable, in_avals):
+def _make_typed_jaxpr(traceable: lu.WrappedFun, in_avals):
   pvals = [pe.PartialVal((aval, core.unit)) for aval in in_avals]
   jaxpr, pvals_out, consts = pe.trace_to_jaxpr(traceable, pvals, instantiate=True)
   out_avals, _ = unzip2(pvals_out)
-  return core.TypedJaxpr(jaxpr, consts, in_avals, out_avals)
+  return core.TypedJaxpr(jaxpr, consts, in_avals, _map(raise_to_shaped, out_avals))
 
 
 def _scan_batching_rule(args, dims, forward, length, jaxpr, num_consts,
@@ -990,7 +1200,7 @@ def _scan_masking_rule(shape_envs, padded_vals, shape_exprs, forward, length,
       *itertools.chain([dynamic_length] + consts, [0], init, xs),
       forward=forward, length=max_length, jaxpr=masked_jaxpr,
       num_consts=1 + num_consts, num_carry=1 + num_carry,
-      linear=[False] + const_linear + [False] + init_linear + xs_linear)
+      linear=tuple([False] + const_linear + [False] + init_linear + xs_linear))
   return out_vals[1:], out_shape
 
 def _masked_scan_jaxpr(jaxpr, num_consts, num_carry):
@@ -1010,25 +1220,18 @@ def _masked_scan_jaxpr(jaxpr, num_consts, num_carry):
   const_avals, carry_avals, x_avals = split_list(jaxpr.in_avals, [num_consts, num_carry])
   return _make_typed_jaxpr(masked, [aval] + const_avals + [aval] + carry_avals + x_avals)
 
-def scan_bind(*args, **kwargs):
-  forward, length, num_consts, num_carry, jaxpr, linear = split_dict(
-      kwargs, ["forward", "length", "num_consts", "num_carry", "jaxpr", "linear"])
-  consts, init, xs = split_list(args, [num_consts, num_carry])
-  assert len(linear) == len(args)
-
-  # check that args match input types
-  consts_avals, init_avals, x_avals = split_list(jaxpr.in_avals, [num_consts, num_carry])
-  xs_avals = _map(partial(_promote_aval_rank, length), x_avals)
-  assert all(_map(typecheck, consts_avals, consts)), (consts, consts_avals)
-  assert all(_map(typecheck, init_avals, init))
-  # assert all(_map(typecheck, xs_avals, xs))
-  # check that output carry type matches input carry type
-  carry_avals, _ = split_list(jaxpr.out_avals, [num_carry])
-  assert all(_map(typematch, init_avals, carry_avals))
-
-  # check that the data flow is sensible
-  core.check_jaxpr(jaxpr.jaxpr)
-
+def scan_bind(*args, forward, length, num_consts, num_carry, jaxpr, linear):
+  if not core.skip_checks:
+    assert len(linear) == len(args)
+    consts, init, xs = split_list(args, [num_consts, num_carry])
+    consts_avals, init_avals, x_avals = split_list(jaxpr.in_avals, [num_consts, num_carry])
+    xs_avals = _map(partial(_promote_aval_rank, length), x_avals)
+    assert all(_map(typecheck, consts_avals, consts)), (consts, consts_avals)
+    assert all(_map(typecheck, init_avals, init))
+    # assert all(_map(typecheck, xs_avals, xs))
+    carry_avals, _ = split_list(jaxpr.out_avals, [num_carry])
+    assert all(_map(typematch, init_avals, carry_avals))
+    core.check_jaxpr(jaxpr.jaxpr)
   return core.Primitive.bind(scan_p, *args, forward=forward, length=length,
                              jaxpr=jaxpr, num_consts=num_consts,
                              num_carry=num_carry, linear=linear)
@@ -1037,10 +1240,13 @@ scan_p = core.Primitive("scan")
 scan_p.multiple_results = True
 scan_p.def_custom_bind(scan_bind)
 scan_p.def_impl(_scan_impl)
+# scan_p.def_impl(partial(xla.apply_primitive, scan_p))  # TODO(mattjj): re-enable
+scan_p.def_abstract_eval(_scan_abstract_eval)
 ad.primitive_jvps[scan_p] = _scan_jvp
 ad.primitive_transposes[scan_p] = _scan_transpose
 pe.custom_partial_eval_rules[scan_p] = _scan_partial_eval
-xla.initial_style_translations[scan_p] = xla.lower_fun(_scan_impl, initial_style=True)
+xla.initial_style_translations[scan_p] = \
+    xla.lower_fun_initial_style(_scan_impl)
 batching.primitive_batchers[scan_p] = _scan_batching_rule
 masking.shape_parameterized_primitive_rules[scan_p] = _scan_masking_rule
 
@@ -1076,8 +1282,7 @@ def map(f, xs):
   return ys
 
 
-def _concat_masking_rule(padded_vals, logical_shapes, dimension, operand_shapes):
-  del operand_shapes  # Unused.
+def _concat_masking_rule(padded_vals, logical_shapes, dimension):
   result = lax.concatenate(padded_vals, dimension)  # fragmented
   offset = 0
   for padded_val, logical_shape in zip(padded_vals, logical_shapes):
@@ -1105,9 +1310,9 @@ def _check_tree(func_name, expected_name, actual_tree, expected_tree):
 def _check_tree_and_avals(what, tree1, avals1, tree2, avals2):
   """Raises TypeError if (tree1, avals1) does not match (tree2, avals2).
 
-  Corresponding `tree` and `avals` must match in the sense that the number of leaves in
-  `tree` must be equal to the length of `avals`.
-  `what` will be prepended to details of the mismatch in TypeError.
+  Corresponding `tree` and `avals` must match in the sense that the number of
+  leaves in `tree` must be equal to the length of `avals`. `what` will be
+  prepended to details of the mismatch in TypeError.
   """
   if tree1 != tree2:
     msg = ("{} must have same type structure, got {} and {}.")
@@ -1188,7 +1393,7 @@ def custom_root(f, initial_guess, solve, tangent_solve):
   _check_tree("solve", "initial_guess", solution_tree, in_tree)
 
   def linearize_and_solve(x, b):
-    unchecked_zeros, f_jvp = api.linearize(f, x)
+    unchecked_zeros, f_jvp = jax.linearize(f, x)
     return tangent_solve(f_jvp, b)
 
   l_and_s_jaxpr, l_and_s_consts, out_tree = _initial_style_jaxpr(
@@ -1249,8 +1454,7 @@ root_p.multiple_results = True
 root_p.def_impl(_root_impl)
 root_p.def_abstract_eval(_root_abstract_eval)
 ad.primitive_jvps[root_p] = _root_jvp
-xla.initial_style_translations[root_p] = xla.lower_fun(
-    _root_impl, initial_style=True)
+xla.initial_style_translations[root_p] = xla.lower_fun_initial_style(_root_impl)
 # TODO(shoyer): write batching rule
 
 
@@ -1271,7 +1475,7 @@ def _transpose_function(linear_fun, primals):
   # TODO(shoyer): can we use something more direct than the vjp machinery?
   # It's particularly awkward that we need the second argument to give
   # particular values of the primals, which are entirely arbitrary.
-  _, vjp_fun = api.vjp(linear_fun, primals)
+  _, vjp_fun = jax.vjp(linear_fun, primals)
 
   def transposed_fun(x):
     (y,) = vjp_fun(x)
@@ -1434,7 +1638,7 @@ def _linear_solve_transpose_rule(cotangent, *primals, **kwargs):
                     'differentiation of custom_linear_solve')
 
   params, b = _split_linear_solve_args(primals, const_lengths)
-  assert b == [ad.undefined_primal] * len(b)
+  assert all(ad.is_undefined_primal(x) for x in b)
   cotangent_b = linear_solve_p.bind(
       *(_flatten(params.transpose()) + cotangent),
       const_lengths=const_lengths.transpose(), jaxprs=jaxprs.transpose(),
@@ -1442,12 +1646,86 @@ def _linear_solve_transpose_rule(cotangent, *primals, **kwargs):
   return [None] * sum(const_lengths) + cotangent_b
 
 
+def _linear_solve_batching_rule(args, dims, **kwargs):
+  const_lengths, jaxprs, tree = split_dict(kwargs,
+                                           ["const_lengths", "jaxprs", "tree"])
+  orig_bat = [d is not batching.not_mapped for d in dims]
+  size, = {
+      a.shape[d] for a, d in zip(args, dims) if d is not batching.not_mapped
+  }
+
+  params, b = _split_linear_solve_args(args, const_lengths)
+  params_dims, b_dims = _split_linear_solve_args(dims, const_lengths)
+  params_bat, orig_b_bat = _split_linear_solve_args(orig_bat, const_lengths)
+
+  (matvec, vecmat, solve, solve_t) = jaxprs
+  (matvec_bat, vecmat_bat, solve_bat, solve_t_bat) = params_bat
+
+  # Fixpoint computation of which parts of x and b are batched; we need to
+  # ensure this is consistent between all four jaxprs
+  b_bat = orig_b_bat
+  x_bat = [False] * len(solve.out_avals)
+  for i in range(1 + len(orig_b_bat) + len(solve.out_avals)):
+    # Apply vecmat and solve -> new batched parts of x
+    solve_jaxpr_batched, solve_x_bat = batching.batch_jaxpr(
+        solve, size, solve_bat + b_bat, instantiate=x_bat)
+    if vecmat is None:
+      vecmat_jaxpr_batched = None
+      x_bat_out = solve_x_bat
+    else:
+      vecmat_jaxpr_batched, vecmat_x_bat = batching.batch_jaxpr(
+          vecmat, size, vecmat_bat + b_bat, instantiate=x_bat)
+      x_bat_out = _map(operator.or_, vecmat_x_bat, solve_x_bat)
+    # Apply matvec and solve_t -> new batched parts of b
+    matvec_jaxpr_batched, matvec_b_bat = batching.batch_jaxpr(
+        matvec, size, matvec_bat + x_bat_out, instantiate=b_bat)
+    if solve_t is None:
+      solve_t_jaxpr_batched = None
+      b_bat_out = _map(operator.or_, matvec_b_bat, orig_b_bat)
+    else:
+      solve_t_jaxpr_batched, solve_t_b_bat = batching.batch_jaxpr(
+          solve_t, size, solve_t_bat + x_bat_out, instantiate=b_bat)
+      b_bat_out = _map(lambda m, s, o: m or s or o, matvec_b_bat, solve_t_b_bat,
+                      orig_b_bat)
+    if x_bat_out == x_bat and b_bat_out == b_bat:
+      break
+    else:
+      x_bat = x_bat_out
+      b_bat = b_bat_out
+  else:
+    assert False, "Fixedpoint not reached"
+
+  batched_jaxprs = _LinearSolveTuple(matvec_jaxpr_batched, vecmat_jaxpr_batched,
+                                     solve_jaxpr_batched, solve_t_jaxpr_batched)
+
+  # Move batched axes to the front
+  new_params = [
+      batching.moveaxis(x, d, 0)
+      if d is not batching.not_mapped and d != 0 else x
+      for x, d in zip(_flatten(params), _flatten(params_dims))
+  ]
+  # Broadcast out b if necessary
+  new_b = [
+      batching.broadcast(x, size, 0) if now_bat and not was_bat else
+      batching.moveaxis(x, d, 0) if now_bat and d != 0 else x
+      for x, d, was_bat, now_bat in zip(b, b_dims, orig_b_bat, b_bat)
+  ]
+
+  outs = linear_solve_p.bind(
+      *(new_params + new_b),
+      const_lengths=const_lengths,
+      jaxprs=batched_jaxprs,
+      tree=tree)
+  out_dims = [0 if batched else batching.not_mapped for batched in b_bat]
+  return outs, out_dims
+
+
 linear_solve_p = core.Primitive('custom_linear_solve')
 linear_solve_p.multiple_results = True
 linear_solve_p.def_impl(_custom_linear_solve_impl)
 linear_solve_p.def_abstract_eval(_linear_solve_abstract_eval)
 ad.primitive_jvps[linear_solve_p] = _custom_linear_solve_jvp
-xla.initial_style_translations[linear_solve_p] = xla.lower_fun(
-    _custom_linear_solve_impl, initial_style=True)
+xla.initial_style_translations[linear_solve_p] = \
+    xla.lower_fun_initial_style(_custom_linear_solve_impl)
 ad.primitive_transposes[linear_solve_p] = _linear_solve_transpose_rule
-# TODO(shoyer): write batching rule
+batching.primitive_batchers[linear_solve_p] = _linear_solve_batching_rule
