@@ -19,12 +19,10 @@ and provide some automatic type mapping logic for converting between Numpy and
 XLA. There are also a handful of related casting utilities.
 """
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
 
 from functools import partial
 import os
+from typing import Callable, Dict
 import warnings
 
 from absl import logging
@@ -59,36 +57,42 @@ flags.DEFINE_string(
     'pass "cpu" for CPU or "gpu" for GPU.')
 
 
-def get_compile_options(num_replicas=None, device_assignment=None):
+def get_compile_options(num_replicas, num_partitions, device_assignment=None):
   """Returns the compile options to use, as derived from flag values.
 
   Args:
-    num_replicas: Optional int indicating the number of replicas for which to
-      compile (default inherited from xla_client.CompileOptions).
+    num_replicas: int indicating the number of replicas for which to compile.
+    num_partitions: int indicating the number of partitions for which to compile.
     device_assignment: Optional tuple of integers indicating the assignment of
       logical replicas to physical devices (default inherited from
-      xla_client.CompileOptions). Must be consistent with `num_replicas`.
+      xla_client.CompileOptions). Must be consistent with `num_replicas` and
+      `num_partitions`.
   """
-  compile_options = None
-  if num_replicas is not None:
-    compile_options = compile_options or xla_client.CompileOptions()
-    compile_options.num_replicas = num_replicas
+  compile_options = xla_client.CompileOptions()
+  compile_options.num_replicas = num_replicas
+  compile_options.num_partitions = num_partitions
   if device_assignment is not None:
-    logging.vlog(2, "get_compile_options: num_replicas=%s device_assignment=%s",
-                 num_replicas, device_assignment)
-    # NOTE(mattjj): xla_client.DeviceAssignment.create expects a 2D ndarray
-    # indexed by replica number and computation per replica, respectively, while
-    # here we currently assume only one computation per replica, hence the
-    # second axis is always trivial.
-    if num_replicas is not None and num_replicas != len(device_assignment):
-      msg = "device_assignment does not match num_replicas: {} vs {}."
-      raise ValueError(msg.format(device_assignment, num_replicas))
-    compile_options = compile_options or xla_client.CompileOptions()
+    logging.vlog(
+        2,
+        'get_compile_options: num_replicas=%s num_partitions=%s device_assignment=%s',
+        num_replicas, num_partitions, device_assignment)
     device_assignment = onp.array(device_assignment)
-    if device_assignment.ndim == 1:
+
+    # Allow 1D device assignment if num_partitions is 1.
+    if (device_assignment.ndim == 1) and (num_partitions == 1):
       device_assignment = device_assignment[:, None]
+
+    if num_replicas != device_assignment.shape[0]:
+      msg = 'device_assignment does not match num_replicas: {} vs {}.'
+      raise ValueError(msg.format(device_assignment, num_replicas))
+
+    if num_partitions != device_assignment.shape[1]:
+      msg = 'device_assignment does not match num_partitions: {} vs {}.'
+      raise ValueError(msg.format(device_assignment, num_partitions))
+
     device_assignment = xla_client.DeviceAssignment.create(device_assignment)
-    assert num_replicas is None or device_assignment.replica_count() == num_replicas
+    assert device_assignment.replica_count() == num_replicas
+    assert device_assignment.computation_count() == num_partitions
     compile_options.device_assignment = device_assignment
   return compile_options
 
@@ -229,8 +233,8 @@ def host_id(backend=None):
 
 
 def host_ids(backend=None):
-  """Returns a list of all host IDs."""
-  return list(set(d.host_id for d in devices(backend)))
+  """Returns a sorted list of all host IDs."""
+  return sorted(list(set(d.host_id for d in devices(backend))))
 
 
 def host_count(backend=None):
@@ -326,7 +330,7 @@ def make_computation_builder(name):
 
 def register_constant_handler(type_, handler_fun):
   _constant_handlers[type_] = handler_fun
-_constant_handlers = {}
+_constant_handlers: Dict[type, Callable] = {}
 
 
 def _ndarray_constant_handler(c, val, canonicalize_types=True):
