@@ -17,8 +17,6 @@ Common neural network layer initializers, consistent with definitions
 used in Keras and Sonnet.
 """
 
-from __future__ import absolute_import
-from __future__ import division
 
 from functools import partial
 
@@ -32,13 +30,13 @@ from jax import random
 def zeros(key, shape, dtype=np.float32): return np.zeros(shape, dtype)
 def ones(key, shape, dtype=np.float32): return np.ones(shape, dtype)
 
-def uniform(scale=1e-2):
-  def init(key, shape, dtype=np.float32):
+def uniform(scale=1e-2, dtype=np.float32):
+  def init(key, shape, dtype=dtype):
     return random.uniform(key, shape, dtype) * scale
   return init
 
-def normal(stddev=1e-2):
-  def init(key, shape, dtype=np.float32):
+def normal(stddev=1e-2, dtype=np.float32):
+  def init(key, shape, dtype=dtype):
     return random.normal(key, shape, dtype) * stddev
   return init
 
@@ -48,8 +46,8 @@ def _compute_fans(shape, in_axis=-2, out_axis=-1):
   fan_out = shape[out_axis] * receptive_field_size
   return fan_in, fan_out
 
-def variance_scaling(scale, mode, distribution, in_axis=-2, out_axis=-1):
-  def init(key, shape, dtype=np.float32):
+def variance_scaling(scale, mode, distribution, in_axis=-2, out_axis=-1, dtype=np.float32):
+  def init(key, shape, dtype=dtype):
     fan_in, fan_out = _compute_fans(shape, in_axis, out_axis)
     if mode == "fan_in": denominator = fan_in
     elif mode == "fan_out": denominator = fan_out
@@ -77,14 +75,14 @@ lecun_normal = partial(variance_scaling, 1.0, "fan_in", "truncated_normal")
 kaiming_uniform = he_uniform = partial(variance_scaling, 2.0, "fan_in", "uniform")
 kaiming_normal = he_normal = partial(variance_scaling, 2.0, "fan_in", "truncated_normal")
 
-def orthogonal(scale=1.0, column_axis=-1):
+def orthogonal(scale=1.0, column_axis=-1, dtype=np.float32):
   """
   Construct an initializer for uniformly distributed orthogonal matrices.
   
   If the shape is not square, the matrices will have orthonormal rows or columns
   depending on which side is smaller.
   """
-  def init(key, shape, dtype=np.float32):
+  def init(key, shape, dtype=dtype):
     if len(shape) < 2:
       raise ValueError("orthogonal initializer requires at least a 2D shape")
     n_rows, n_cols = onp.prod(shape) // shape[column_axis], shape[column_axis]
@@ -99,21 +97,21 @@ def orthogonal(scale=1.0, column_axis=-1):
   return init
 
 
-def delta_orthogonal(scale=1.0, column_axis=-1):
+def delta_orthogonal(scale=1.0, column_axis=-1, dtype=np.float32):
   """
   Construct an initializer for delta orthogonal kernels; see arXiv:1806.05393. 
 
   The shape must be 3D, 4D or 5D.
   """
-  def init(key, shape, dtype=np.float32):
+  def init(key, shape, dtype=dtype):
     if len(shape) not in [3, 4, 5]:
       raise ValueError("Delta orthogonal initializer requires a 3D, 4D or 5D "
                        "shape.")
     if shape[-1] < shape[-2]:
       raise ValueError("`fan_in` must be less or equal than `fan_out`. ")
-    ortho_init = orthogonal(scale=scale, column_axis=column_axis)
+    ortho_init = orthogonal(scale=scale, column_axis=column_axis, dtype=dtype)
     ortho_matrix = ortho_init(key, shape[-2:])
-    W = np.zeros(shape)
+    W = np.zeros(shape, dtype=dtype)
     if len(shape) == 3:
       k = shape[0]
       return ops.index_update(W, ops.index[(k-1)//2, ...], ortho_matrix)
