@@ -20,9 +20,6 @@ The JAX PRNG system is based on "Parallel random numbers: as easy as 1, 2, 3"
 https://github.com/google/jax/blob/master/design_notes/prng.md
 """
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
 
 from functools import partial
 import itertools
@@ -33,7 +30,7 @@ from . import lax
 from . import numpy as np
 from . import tree_util
 from . import dtypes
-from .api import custom_transforms, defjvp, jit, vmap
+from .api import jit, vmap
 from .numpy.lax_numpy import _constant_like, asarray, stack
 from jax.lib import xla_bridge
 from jax.lib import cuda_prng
@@ -195,9 +192,9 @@ threefry2x32_p.def_impl(partial(xla.apply_primitive, threefry2x32_p))
 threefry2x32_p.def_abstract_eval(_threefry2x32_abstract_eval)
 batching.defbroadcasting(threefry2x32_p)
 xla.translations[threefry2x32_p] = xla.lower_fun(
-    partial(_threefry2x32_lowering, use_rolled_loops=False), instantiate=True)
+    partial(_threefry2x32_lowering, use_rolled_loops=False))
 xla.backend_specific_translations['cpu'][threefry2x32_p] = xla.lower_fun(
-    partial(_threefry2x32_lowering, use_rolled_loops=True), instantiate=True)
+    partial(_threefry2x32_lowering, use_rolled_loops=True))
 if cuda_prng:
   xla.backend_specific_translations['gpu'][threefry2x32_p] = \
       _threefry2x32_gpu_translation_rule
@@ -293,9 +290,9 @@ def _random_bits(key, bit_width, shape):
 def _check_shape(name, shape, *param_shapes):
   try:
     shape = tuple(map(int, shape))
-  except TypeError:
+  except TypeError as err:
     msg = "{} requires a concrete tuple of integers as shape argument, got {}."
-    raise ValueError(msg.format(name, shape))
+    raise ValueError(msg.format(name, shape)) from err
   if param_shapes:
     shape_ = lax.broadcast_shapes(shape, *param_shapes)
     if shape != shape_:
@@ -321,6 +318,7 @@ def uniform(key, shape=(), dtype=onp.float64, minval=0., maxval=1.):
     A random array with the specified shape and dtype.
   """
   dtype = dtypes.canonicalize_dtype(dtype)
+  shape = abstract_arrays.canonicalize_shape(shape)
   return _uniform(key, shape, dtype, minval, maxval)
 
 @partial(jit, static_argnums=(1, 2))
@@ -369,6 +367,7 @@ def randint(key, shape, minval, maxval, dtype=onp.int64):
     A random array with the specified shape and dtype.
   """
   dtype = dtypes.canonicalize_dtype(dtype)
+  shape = abstract_arrays.canonicalize_shape(shape)
   return _randint(key, shape, minval, maxval, dtype)
 
 @partial(jit, static_argnums=(1, 4))
@@ -467,6 +466,7 @@ def normal(key, shape=(), dtype=onp.float64):
     A random array with the specified shape and dtype.
   """
   dtype = dtypes.canonicalize_dtype(dtype)
+  shape = abstract_arrays.canonicalize_shape(shape)
   return _normal(key, shape, dtype)
 
 @partial(jit, static_argnums=(1, 2))
@@ -500,6 +500,8 @@ def multivariate_normal(key, mean, cov, shape=None, dtype=onp.float64):
     ``broadcast_shapes(mean.shape[:-1], cov.shape[:-2]) + mean.shape[-1:]``.
   """
   dtype = dtypes.canonicalize_dtype(dtype)
+  if shape is not None:
+    shape = abstract_arrays.canonicalize_shape(shape)
   return _multivariate_normal(key, mean, cov, shape, dtype)
 
 @partial(jit, static_argnums=(3, 4))
@@ -547,6 +549,8 @@ def truncated_normal(key, lower, upper, shape=None, dtype=onp.float64):
     ``shape`` is not None, or else by broadcasting ``lower`` and ``upper``.
   """
   dtype = dtypes.canonicalize_dtype(dtype)
+  if shape is not None:
+    shape = abstract_arrays.canonicalize_shape(shape)
   return _truncated_normal(key, lower, upper, shape, dtype)
 
 @partial(jit, static_argnums=(3, 4))
@@ -581,6 +585,8 @@ def bernoulli(key, p=onp.float32(0.5), shape=None):
     is not None, or else ``p.shape``.
   """
   dtype = dtypes.canonicalize_dtype(lax.dtype(p))
+  if shape is not None:
+    shape = abstract_arrays.canonicalize_shape(shape)
   if not np.issubdtype(dtype, onp.floating):
     msg = "bernoulli probability `p` must have a floating dtype, got {}."
     raise TypeError(msg.format(dtype))
@@ -617,6 +623,8 @@ def beta(key, a, b, shape=None, dtype=onp.float64):
     ``shape`` is not None, or else by broadcasting ``a`` and ``b``.
   """
   dtype = dtypes.canonicalize_dtype(dtype)
+  if shape is not None:
+    shape = abstract_arrays.canonicalize_shape(shape)
   return _beta(key, a, b, shape, dtype)
 
 def _beta(key, a, b, shape, dtype):
@@ -649,6 +657,7 @@ def cauchy(key, shape=(), dtype=onp.float64):
     A random array with the specified shape and dtype.
   """
   dtype = dtypes.canonicalize_dtype(dtype)
+  shape = abstract_arrays.canonicalize_shape(shape)
   return _cauchy(key, shape, dtype)
 
 @partial(jit, static_argnums=(1, 2))
@@ -660,7 +669,7 @@ def _cauchy(key, shape, dtype):
 
 
 def dirichlet(key, alpha, shape=None, dtype=onp.float64):
-  """Sample Cauchy random values with given shape and float dtype.
+  """Sample Dirichlet random values with given shape and float dtype.
 
   Args:
     key: a PRNGKey used as the random key.
@@ -680,6 +689,8 @@ def dirichlet(key, alpha, shape=None, dtype=onp.float64):
     ``alpha.shape``.
   """
   dtype = dtypes.canonicalize_dtype(dtype)
+  if shape is not None:
+    shape = abstract_arrays.canonicalize_shape(shape)
   return _dirichlet(key, alpha, shape, dtype)
 
 @partial(jit, static_argnums=(2, 3))
@@ -712,6 +723,7 @@ def exponential(key, shape=(), dtype=onp.float64):
     A random array with the specified shape and dtype.
   """
   dtype = dtypes.canonicalize_dtype(dtype)
+  shape = abstract_arrays.canonicalize_shape(shape)
   return _exponential(key, shape, dtype)
 
 @partial(jit, static_argnums=(1, 2))
@@ -785,99 +797,99 @@ _bivariate_coef = [[0.16009398, -0.094634816, 0.025146379, -0.0030648348,
                     0.017050642, -0.0021309345, 0.00085092385, -1.5248239e-07]]
 
 def _gamma_grad_one(z, alpha):
-    # Ref 1: Pathwise Derivatives Beyond the Reparameterization Trick, Martin & Fritz
-    # Ref 2: Case 4 follows https://github.com/fritzo/notebooks/blob/master/gamma-reparameterized.ipynb
+  # Ref 1: Pathwise Derivatives Beyond the Reparameterization Trick, Martin & Fritz
+  # Ref 2: Case 4 follows https://github.com/fritzo/notebooks/blob/master/gamma-reparameterized.ipynb
 
-    # TODO: use lax.cond instead of lax.while_loop when its batching rule is available
-    # See https://github.com/google/jax/issues/490
-    def _case1(zagf):
-        z, alpha, _, flag = zagf
+  # TODO: use lax.cond instead of lax.while_loop when its batching rule is available
+  # See https://github.com/google/jax/issues/490
+  def _case1(zagf):
+    z, alpha, _, flag = zagf
 
-        # dz = - dCDF(z; a) / pdf(z; a)
-        # pdf = z^(a-1) * e^(-z) / Gamma(a)
-        # CDF(z; a) = IncompleteGamma(a, z) / Gamma(a)
-        # dCDF(z; a) = (dIncompleteGamma - IncompleteGamma * Digamma(a)) / Gamma(a)
-        #            =: unnormalized_dCDF / Gamma(a)
-        # IncompleteGamma ~ z^a [ 1/a - z/(a+1) + z^2/2!(a+2) - z^3/3!(a+3) + z^4/4!(a+4) - z^5/5!(a+5) ]
-        #                 =: z^a * term1
-        # dIncompleteGamma ~ z^a * log(z) * term1 - z^a [1/a^2 - z/(a+1)^2 + z^2/2!(a+2)^2
-        #                                                - z^3/3!(a+3)^2 + z^4/4!(a+4)^2 - z^5/5!(a+5)^2 ]
-        #                  =: z^a * log(z) * term1 - z^a * term2
-        # unnormalized_dCDF = z^a { [log(z) - Digamma(a)] * term1 - term2 }
-        zi = 1.0
-        update = zi / alpha
-        term1 = update
-        term2 = update / alpha
-        for i in range(1, 6):
-            zi = -zi * z / i
-            update = zi / (alpha + i)
-            term1 = term1 + update
-            term2 = term2 + update / (alpha + i)
+    # dz = - dCDF(z; a) / pdf(z; a)
+    # pdf = z^(a-1) * e^(-z) / Gamma(a)
+    # CDF(z; a) = IncompleteGamma(a, z) / Gamma(a)
+    # dCDF(z; a) = (dIncompleteGamma - IncompleteGamma * Digamma(a)) / Gamma(a)
+    #            =: unnormalized_dCDF / Gamma(a)
+    # IncompleteGamma ~ z^a [ 1/a - z/(a+1) + z^2/2!(a+2) - z^3/3!(a+3) + z^4/4!(a+4) - z^5/5!(a+5) ]
+    #                 =: z^a * term1
+    # dIncompleteGamma ~ z^a * log(z) * term1 - z^a [1/a^2 - z/(a+1)^2 + z^2/2!(a+2)^2
+    #                                                - z^3/3!(a+3)^2 + z^4/4!(a+4)^2 - z^5/5!(a+5)^2 ]
+    #                  =: z^a * log(z) * term1 - z^a * term2
+    # unnormalized_dCDF = z^a { [log(z) - Digamma(a)] * term1 - term2 }
+    zi = 1.0
+    update = zi / alpha
+    term1 = update
+    term2 = update / alpha
+    for i in range(1, 6):
+        zi = -zi * z / i
+        update = zi / (alpha + i)
+        term1 = term1 + update
+        term2 = term2 + update / (alpha + i)
 
-        unnormalized_cdf_dot = np.power(z, alpha) * ((np.log(z) - lax.digamma(alpha)) * term1 - term2)
-        unnormalized_pdf = np.power(z, alpha - 1) * np.exp(-z)
-        grad = -unnormalized_cdf_dot / unnormalized_pdf
+    unnormalized_cdf_dot = np.power(z, alpha) * ((np.log(z) - lax.digamma(alpha)) * term1 - term2)
+    unnormalized_pdf = np.power(z, alpha - 1) * np.exp(-z)
+    grad = -unnormalized_cdf_dot / unnormalized_pdf
 
-        return z, alpha, grad, ~flag
+    return z, alpha, grad, ~flag
 
-    def _cond2(zagf):
-        z, alpha, _, flag = zagf
-        return (~flag) & (alpha > 8.0) & ((z < 0.9 * alpha) | (z > 1.1 * alpha))
+  def _cond2(zagf):
+    z, alpha, _, flag = zagf
+    return (~flag) & (alpha > 8.0) & ((z < 0.9 * alpha) | (z > 1.1 * alpha))
 
-    def _case2(zagf):
-        z, alpha, _, flag = zagf
+  def _case2(zagf):
+    z, alpha, _, flag = zagf
 
-        # Formula 58 of [1]
-        sqrt_8a = np.sqrt(8 * alpha)
-        z_minus_a = z - alpha
-        log_z_div_a = np.log(z / alpha)
-        sign = np.where(z < alpha, 1.0, -1.0)
-        term1 = 4 * (z + alpha) / (sqrt_8a * z_minus_a * z_minus_a)
-        term2 = log_z_div_a * (sqrt_8a / z_minus_a + sign * np.power(z_minus_a - alpha * log_z_div_a, -1.5))
-        term3 = z * (1.0 + 1.0 / (12 * alpha) + 1.0 / (288 * alpha * alpha)) / sqrt_8a
-        grad = (term1 + term2) * term3
+    # Formula 58 of [1]
+    sqrt_8a = np.sqrt(8 * alpha)
+    z_minus_a = z - alpha
+    log_z_div_a = np.log(z / alpha)
+    sign = np.where(z < alpha, lax._const(z, 1.0), lax._const(z, -1.0))
+    term1 = 4 * (z + alpha) / (sqrt_8a * z_minus_a * z_minus_a)
+    term2 = log_z_div_a * (sqrt_8a / z_minus_a + sign * np.power(z_minus_a - alpha * log_z_div_a, -1.5))
+    term3 = z * (1.0 + 1.0 / (12 * alpha) + 1.0 / (288 * alpha * alpha)) / sqrt_8a
+    grad = (term1 + term2) * term3
 
-        return z, alpha, grad, ~flag
+    return z, alpha, grad, ~flag
 
-    def _cond3(zagf):
-        z, alpha, _, flag = zagf
-        return (~flag) & (alpha > 8.0) & (z >= 0.9 * alpha) & (z <= 1.1 * alpha)
+  def _cond3(zagf):
+    z, alpha, _, flag = zagf
+    return (~flag) & (alpha > 8.0) & (z >= 0.9 * alpha) & (z <= 1.1 * alpha)
 
-    def _case3(zagf):
-        z, alpha, _, flag = zagf
+  def _case3(zagf):
+    z, alpha, _, flag = zagf
 
-        # Formula 59 of [1]
-        z_div_a = np.divide(z, alpha)
-        aa = alpha * alpha
-        term1 = 1440 * alpha + 6 * z_div_a * (53 - 120 * z) - 65 * z_div_a * z_div_a + 3600 * z + 107
-        term2 = 1244160 * alpha * aa
-        term3 = 1 + 24 * alpha + 288 * aa
-        grad = term1 * term3 / term2
+    # Formula 59 of [1]
+    z_div_a = np.divide(z, alpha)
+    aa = alpha * alpha
+    term1 = 1440 * alpha + 6 * z_div_a * (53 - 120 * z) - 65 * z_div_a * z_div_a + 3600 * z + 107
+    term2 = 1244160 * alpha * aa
+    term3 = 1 + 24 * alpha + 288 * aa
+    grad = term1 * term3 / term2
 
-        return z, alpha, grad, ~flag
+    return z, alpha, grad, ~flag
 
-    def _case4(zagf):
-        z, alpha, _, flag = zagf
+  def _case4(zagf):
+    z, alpha, _, flag = zagf
 
-        # Ref [2]
-        u = np.log(z / alpha)
-        v = np.log(alpha)
-        c = []
-        for i in range(8):
-            c.append(_bivariate_coef[0][i] + u * (_bivariate_coef[1][i] + u * _bivariate_coef[2][i]))
-        p = c[0] + v * (c[1] + v * (c[2] + v * c[3]))
-        q = c[4] + v * (c[5] + v * (c[6] + v * c[7]))
-        grad = np.exp(p / np.maximum(q, 0.01))
+    # Ref [2]
+    u = np.log(z / alpha)
+    v = np.log(alpha)
+    c = []
+    for i in range(8):
+        c.append(_bivariate_coef[0][i] + u * (_bivariate_coef[1][i] + u * _bivariate_coef[2][i]))
+    p = c[0] + v * (c[1] + v * (c[2] + v * c[3]))
+    q = c[4] + v * (c[5] + v * (c[6] + v * c[7]))
+    grad = np.exp(p / np.maximum(q, 0.01))
 
-        return z, alpha, grad, ~flag
+    return z, alpha, grad, ~flag
 
-    _, _, grad, flag = lax.while_loop(lambda zagf: (~zagf[3]) & (zagf[0] < 0.8),
-                                      _case1,
-                                      (z, alpha, lax._const(alpha, 0.0), False))
-    _, _, grad, flag = lax.while_loop(_cond2, _case2, (z, alpha, grad, flag))
-    _, _, grad, flag = lax.while_loop(_cond3, _case3, (z, alpha, grad, flag))
-    _, _, grad, flag = lax.while_loop(lambda zagf: ~zagf[3], _case4, (z, alpha, grad, flag))
-    return grad
+  _, _, grad, flag = lax.while_loop(lambda zagf: (~zagf[3]) & (zagf[0] < 0.8),
+                                    _case1,
+                                    (z, alpha, lax._const(alpha, 0.0), False))
+  _, _, grad, flag = lax.while_loop(_cond2, _case2, (z, alpha, grad, flag))
+  _, _, grad, flag = lax.while_loop(_cond3, _case3, (z, alpha, grad, flag))
+  _, _, grad, flag = lax.while_loop(lambda zagf: ~zagf[3], _case4, (z, alpha, grad, flag))
+  return grad
 
 def _gamma_grad(sample, a):
   samples = np.reshape(sample, -1)
@@ -915,7 +927,7 @@ random_gamma_p.multiple_results = True
 random_gamma_p.def_impl(_gamma_impl)
 random_gamma_p.def_abstract_eval(lambda key, a: (abstract_arrays.raise_to_shaped(a),))
 ad.defjvp2(random_gamma_p, None, lambda tangent, ans, key, a: (tangent * _gamma_grad(ans[0], a),))
-xla.translations[random_gamma_p] = xla.lower_fun(_gamma_impl, instantiate=True)
+xla.translations[random_gamma_p] = xla.lower_fun(_gamma_impl)
 batching.primitive_batchers[random_gamma_p] = _gamma_batching_rule
 
 def gamma(key, a, shape=None, dtype=onp.float64):
@@ -936,6 +948,8 @@ def gamma(key, a, shape=None, dtype=onp.float64):
     ``shape`` is not None, or else by ``a.shape``.
   """
   dtype = dtypes.canonicalize_dtype(dtype)
+  if shape is not None:
+    shape = abstract_arrays.canonicalize_shape(shape)
   return _gamma(key, a, shape, dtype)
 
 @partial(jit, static_argnums=(2, 3))
@@ -965,6 +979,7 @@ def gumbel(key, shape=(), dtype=onp.float64):
     A random array with the specified shape and dtype.
   """
   dtype = dtypes.canonicalize_dtype(dtype)
+  shape = abstract_arrays.canonicalize_shape(shape)
   return _gumbel(key, shape, dtype)
 
 @partial(jit, static_argnums=(1, 2))
@@ -1016,6 +1031,7 @@ def laplace(key, shape=(), dtype=onp.float64):
     A random array with the specified shape and dtype.
   """
   dtype = dtypes.canonicalize_dtype(dtype)
+  shape = abstract_arrays.canonicalize_shape(shape)
   return _laplace(key, shape, dtype)
 
 @partial(jit, static_argnums=(1, 2))
@@ -1040,6 +1056,7 @@ def logistic(key, shape=(), dtype=onp.float64):
     A random array with the specified shape and dtype.
   """
   dtype = dtypes.canonicalize_dtype(dtype)
+  shape = abstract_arrays.canonicalize_shape(shape)
   return _logistic(key, shape, dtype)
 
 @partial(jit, static_argnums=(1, 2))
@@ -1066,6 +1083,8 @@ def pareto(key, b, shape=None, dtype=onp.float64):
     ``shape`` is not None, or else by ``b.shape``.
   """
   dtype = dtypes.canonicalize_dtype(dtype)
+  if shape is not None:
+    shape = abstract_arrays.canonicalize_shape(shape)
   return _pareto(key, b, shape, dtype)
 
 @partial(jit, static_argnums=(2, 3))
@@ -1098,6 +1117,7 @@ def t(key, df, shape=(), dtype=onp.float64):
     ``shape`` is not None, or else by ``df.shape``.
   """
   dtype = dtypes.canonicalize_dtype(dtype)
+  shape = abstract_arrays.canonicalize_shape(shape)
   return _t(key, df, shape, dtype)
 
 @partial(jit, static_argnums=(2, 3))
