@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import absolute_import
 from . import bernoulli
 from . import poisson
 from . import beta
@@ -26,3 +25,4 @@ from . import norm
 from . import pareto
 from . import t
 from . import uniform
+from . import logistic
