@@ -12,11 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
 
-import collections
 import functools
 import itertools as it
 import types
@@ -228,3 +224,9 @@ def get_module_functions(module):
         attr, (types.BuiltinFunctionType, types.FunctionType, onp.ufunc)):
       module_fns.add(attr)
   return module_fns
+
+def wrap_name(name, transform_name):
+  return transform_name + '(' + name + ')'
+
+def extend_name_stack(stack, name=''):
+  return stack + name + '/'
