@@ -13,13 +13,11 @@
 # limitations under the License.
 
 """Tests for the LAPAX linear algebra module."""
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
 
 from functools import partial
 import itertools
 import unittest
+import sys
 
 import numpy as onp
 import scipy as osp
@@ -54,6 +52,10 @@ def _skip_if_unsupported_type(dtype):
       dtype in (onp.dtype('float64'), onp.dtype('complex128'))):
     raise unittest.SkipTest("--jax_enable_x64 is not set")
 
+# TODO(phawkins): bug https://github.com/google/jax/issues/432
+def _skip_on_mac_xla_bug():
+  if sys.platform == "darwin" and osp.version.version > "1.0.0":
+    raise unittest.SkipTest("Test fails on Mac with new scipy (issue #432)")
 
 class NumpyLinalgTest(jtu.JaxTestCase):
 
@@ -73,8 +75,7 @@ class NumpyLinalgTest(jtu.JaxTestCase):
       return [onp.matmul(a, np.conj(T(a)))]
 
     if (np.issubdtype(dtype, np.complexfloating) and
-        (jtu.device_under_test() == "tpu" or
-         (jtu.device_under_test() == "cpu" and jax.lib.version < (0, 1, 38)))):
+        jtu.device_under_test() == "tpu"):
       self.skipTest("Unimplemented case for complex Cholesky decomposition.")
 
     self._CheckAgainstNumpy(onp.linalg.cholesky, np.linalg.cholesky, args_maker,
@@ -114,6 +115,47 @@ class NumpyLinalgTest(jtu.JaxTestCase):
 
   @parameterized.named_parameters(jtu.cases_from_list(
       {"testcase_name":
+       "_m={}_n={}_q={}".format(
+            jtu.format_shape_dtype_string((m,), dtype),
+            jtu.format_shape_dtype_string((nq[0],), dtype),
+            jtu.format_shape_dtype_string(nq[1], dtype)),
+       "m": m,
+       "nq": nq, "dtype": dtype, "rng_factory": rng_factory}
+      for m in [1, 5, 7, 23]
+      for nq in zip([2, 4, 6, 36], [(1, 2), (2, 2), (1, 2, 3), (3, 3, 1, 4)])
+      for dtype in float_types
+      for rng_factory in [jtu.rand_default]))
+  def testTensorsolve(self, m, nq, dtype, rng_factory):
+    rng = rng_factory()
+    _skip_if_unsupported_type(dtype)
+    if m == 23:
+      _skip_on_mac_xla_bug()
+    
+    # According to numpy docs the shapes are as follows:
+    # Coefficient tensor (a), of shape b.shape + Q. 
+    # And prod(Q) == prod(b.shape) 
+    # Therefore, n = prod(q) 
+    n, q = nq
+    b_shape = (n, m)
+    # To accomplish prod(Q) == prod(b.shape) we append the m extra dim
+    # to Q shape
+    Q = q + (m,)
+    args_maker = lambda: [
+        rng(b_shape + Q, dtype), # = a
+        rng(b_shape, dtype)]     # = b
+    a, b = args_maker()
+    result = np.linalg.tensorsolve(*args_maker())
+    self.assertEqual(result.shape, Q)
+
+    self._CheckAgainstNumpy(onp.linalg.tensorsolve, 
+                            np.linalg.tensorsolve, args_maker,
+                            check_dtypes=True, tol=1e-3)
+    self._CompileAndCheck(np.linalg.tensorsolve, 
+                          args_maker, check_dtypes=True,
+                          rtol={onp.float64: 1e-13})
+
+  @parameterized.named_parameters(jtu.cases_from_list(
+      {"testcase_name":
        "_shape={}".format(jtu.format_shape_dtype_string(shape, dtype)),
        "shape": shape, "dtype": dtype, "rng_factory": rng_factory}
       for shape in [(0, 0), (1, 1), (3, 3), (4, 4), (10, 10), (200, 200),
@@ -138,6 +180,7 @@ class NumpyLinalgTest(jtu.JaxTestCase):
       for dtype in float_types
       for rng_factory in [jtu.rand_default]))
   @jtu.skip_on_devices("tpu")
+  @jtu.skip_on_flag("jax_skip_slow_tests", True)
   def testSlogdetGrad(self, shape, dtype, rng_factory):
     rng = rng_factory()
     _skip_if_unsupported_type(dtype)
@@ -192,6 +235,8 @@ class NumpyLinalgTest(jtu.JaxTestCase):
   def testEigvals(self, shape, dtype, rng_factory):
     rng = rng_factory()
     _skip_if_unsupported_type(dtype)
+    if shape == (50, 50) and dtype == onp.complex64:
+      _skip_on_mac_xla_bug()
     n = shape[-1]
     args_maker = lambda: [rng(shape, dtype)]
     a, = args_maker()
@@ -388,7 +433,7 @@ class NumpyLinalgTest(jtu.JaxTestCase):
              (isinstance(axis, tuple) and len(axis) == 1)
           else [None, 'fro', 1, 2, -1, -2, np.inf, -np.inf, 'nuc'])
       for dtype in float_types + complex_types
-      for rng_factory in [jtu.rand_default]))
+      for rng_factory in [jtu.rand_default]))  # type: ignore
   def testNorm(self, shape, dtype, ord, axis, keepdims, rng_factory):
     rng = rng_factory()
     _skip_if_unsupported_type(dtype)
@@ -456,8 +501,8 @@ class NumpyLinalgTest(jtu.JaxTestCase):
 
     self._CompileAndCheck(partial(np.linalg.svd, full_matrices=full_matrices, compute_uv=compute_uv),
                           args_maker, check_dtypes=True)
-    if not full_matrices:
-      svd = partial(np.linalg.svd, full_matrices=False)
+    if not (compute_uv and full_matrices):
+      svd = partial(np.linalg.svd, full_matrices=full_matrices, compute_uv=compute_uv)
       jtu.check_jvp(svd, partial(jvp, svd), (a,), rtol=1e-2, atol=1e-1)
 
   @parameterized.named_parameters(jtu.cases_from_list(
@@ -536,6 +581,69 @@ class NumpyLinalgTest(jtu.JaxTestCase):
 
   @parameterized.named_parameters(jtu.cases_from_list(
       {"testcase_name":
+       "_shape={}_pnorm={}".format(jtu.format_shape_dtype_string(shape, dtype), pnorm),
+       "shape": shape, "pnorm": pnorm, "dtype": dtype}
+      for shape in [(1, 1), (4, 4), (2, 3, 5), (5, 5, 5), (20, 20), (5, 10)]
+      for pnorm in [np.inf, -np.inf, 1, -1, 2, -2, 'fro']
+      for dtype in float_types + complex_types))
+  @jtu.skip_on_devices("tpu")  # SVD is not implemented on the TPU backend
+  @jtu.skip_on_devices("gpu")  # TODO(#2203): numerical errors
+  def testCond(self, shape, pnorm, dtype):
+    _skip_if_unsupported_type(dtype)
+
+    def gen_mat():
+      arr_gen = jtu.rand_some_nan()
+      res = arr_gen(shape, dtype)
+      return res
+
+    def args_gen(p):
+      def _args_gen():
+        return [gen_mat(), p]
+      return _args_gen
+
+    args_maker = args_gen(pnorm)
+    if pnorm not in [2, -2] and len(set(shape[-2:])) != 1:
+      with self.assertRaises(onp.linalg.LinAlgError):
+        np.linalg.cond(*args_maker())
+    else:
+      self._CheckAgainstNumpy(onp.linalg.cond, np.linalg.cond, args_maker,
+                              check_dtypes=False, tol=1e-3)
+      partial_norm = partial(np.linalg.cond, p=pnorm)
+      self._CompileAndCheck(partial_norm, lambda: [gen_mat()],
+                            check_dtypes=False, rtol=1e-03, atol=1e-03)
+
+  @parameterized.named_parameters(jtu.cases_from_list(
+      {"testcase_name":
+       "_shape={}".format(jtu.format_shape_dtype_string(shape, dtype)),
+       "shape": shape, "dtype": dtype, "rng_factory": rng_factory}
+      for shape in [(1, 1), (4, 4), (200, 200), (7, 7, 7, 7)]
+      for dtype in float_types
+      for rng_factory in [jtu.rand_default]))
+  def testTensorinv(self, shape, dtype, rng_factory):
+    _skip_if_unsupported_type(dtype)
+    if shape[0] > 100:
+      _skip_on_mac_xla_bug()
+    rng = rng_factory()
+
+    def tensor_maker():
+      invertible = False
+      while not invertible:
+        a = rng(shape, dtype)
+        try:
+          onp.linalg.inv(a)
+          invertible = True
+        except onp.linalg.LinAlgError:
+          pass
+      return a
+
+    args_maker = lambda: [tensor_maker(), int(onp.floor(len(shape) / 2))]
+    self._CheckAgainstNumpy(onp.linalg.tensorinv, np.linalg.tensorinv, args_maker,
+                            check_dtypes=False, tol=1e-3)
+    partial_inv = partial(np.linalg.tensorinv, ind=int(onp.floor(len(shape) / 2)))
+    self._CompileAndCheck(partial_inv, lambda: [tensor_maker()], check_dtypes=False, rtol=1e-03, atol=1e-03)
+
+  @parameterized.named_parameters(jtu.cases_from_list(
+      {"testcase_name":
        "_lhs={}_rhs={}".format(
            jtu.format_shape_dtype_string(lhs_shape, dtype),
            jtu.format_shape_dtype_string(rhs_shape, dtype)),
@@ -569,6 +677,8 @@ class NumpyLinalgTest(jtu.JaxTestCase):
   def testInv(self, shape, dtype, rng_factory):
     rng = rng_factory()
     _skip_if_unsupported_type(dtype)
+    if shape == (200, 200) and dtype == onp.float32:
+      _skip_on_mac_xla_bug()
     if jtu.device_under_test() == "gpu" and shape == (200, 200):
       raise unittest.SkipTest("Test is flaky on GPU")
 
@@ -598,11 +708,53 @@ class NumpyLinalgTest(jtu.JaxTestCase):
   def testPinv(self, shape, dtype, rng_factory):
     rng = rng_factory()
     _skip_if_unsupported_type(dtype)
+    if shape == (7, 10000) and dtype in [onp.complex64, onp.float32]:
+      _skip_on_mac_xla_bug()
     args_maker = lambda: [rng(shape, dtype)]
 
     self._CheckAgainstNumpy(onp.linalg.pinv, np.linalg.pinv, args_maker,
                             check_dtypes=True, tol=1e-3)
     self._CompileAndCheck(np.linalg.pinv, args_maker, check_dtypes=True)
+
+  @parameterized.named_parameters(jtu.cases_from_list(
+      {"testcase_name": "_shape={}_n={}".format(
+          jtu.format_shape_dtype_string(shape, dtype), n),
+       "shape": shape, "dtype": dtype, "n": n, "rng_factory": rng_factory}
+      for shape in [(1, 1), (2, 2), (4, 4), (5, 5),
+                    (1, 2, 2), (2, 3, 3), (2, 5, 5)]
+      for dtype in float_types + complex_types
+      for n in [-5, -2, -1, 0, 1, 2, 3, 4, 5, 10]
+      for rng_factory in [jtu.rand_default]))
+  @jtu.skip_on_devices("tpu")  # TODO(b/149870255): Bug in XLA:TPU?.
+  def testMatrixPower(self, shape, dtype, n, rng_factory):
+    rng = rng_factory()
+    _skip_if_unsupported_type(dtype)
+    args_maker = lambda: [rng(shape, dtype)]
+    tol = 1e-1 if jtu.device_under_test() == "tpu" else 1e-3
+    self._CheckAgainstNumpy(partial(onp.linalg.matrix_power, n=n),
+                            partial(np.linalg.matrix_power, n=n),
+                            args_maker, check_dtypes=True, tol=tol)
+    self._CompileAndCheck(partial(np.linalg.matrix_power, n=n), args_maker,
+                          check_dtypes=True, rtol=1e-3)
+
+  @parameterized.named_parameters(jtu.cases_from_list(
+      {"testcase_name": "_shape={}".format(
+           jtu.format_shape_dtype_string(shape, dtype)),
+       "shape": shape, "dtype": dtype, "rng_factory": rng_factory}
+      for shape in [(3, ), (1, 2), (8, 5), (4, 4), (5, 5), (50, 50)]
+      for dtype in float_types + complex_types
+      for rng_factory in [jtu.rand_default]))
+  @jtu.skip_on_devices("gpu", "tpu")  # TODO(b/145608614): SVD crashes on GPU.
+  def testMatrixRank(self, shape, dtype, rng_factory):
+    rng = rng_factory()
+    _skip_if_unsupported_type(dtype)
+    n = shape[-1]
+    args_maker = lambda: [rng(shape, dtype)]
+    a, = args_maker()
+    self._CheckAgainstNumpy(onp.linalg.matrix_rank, np.linalg.matrix_rank,
+                            args_maker, check_dtypes=False, tol=1e-3)
+    self._CompileAndCheck(np.linalg.matrix_rank, args_maker,
+                          check_dtypes=False, rtol=1e-3)
 
   # Regression test for incorrect type for eigenvalues of a complex matrix.
   @jtu.skip_on_devices("tpu")  # TODO(phawkins): No complex eigh implementation on TPU.
@@ -615,6 +767,7 @@ class NumpyLinalgTest(jtu.JaxTestCase):
     xc = onp.eye(3, dtype=onp.complex)
     self.assertAllClose(xc, grad_test_jc(xc), check_dtypes=True)
 
+  @jtu.skip_on_flag("jax_skip_slow_tests", True)
   def testIssue1151(self):
     A = np.array(onp.random.randn(100, 3, 3), dtype=np.float32)
     b = np.array(onp.random.randn(100, 3), dtype=np.float32)
@@ -626,6 +779,7 @@ class NumpyLinalgTest(jtu.JaxTestCase):
     jac0 = jax.jacobian(np.linalg.solve, argnums=0)(A[0], b[0])
     jac1 = jax.jacobian(np.linalg.solve, argnums=1)(A[0], b[0])
 
+  @jtu.skip_on_flag("jax_skip_slow_tests", True)
   def testIssue1383(self):
     seed = jax.random.PRNGKey(0)
     tmp = jax.random.uniform(seed, (2,2))
@@ -642,6 +796,23 @@ class NumpyLinalgTest(jtu.JaxTestCase):
 
 
 class ScipyLinalgTest(jtu.JaxTestCase):
+
+  @parameterized.named_parameters(jtu.cases_from_list(
+      {"testcase_name": "_i={}".format(i), "args": args}
+      for i, args in enumerate([
+        (),
+        (1,),
+        (7, -2),
+        (3, 4, 5),
+        (onp.ones((3, 4), dtype=np.float_), 5,
+         onp.random.randn(5, 2).astype(np.float_)),
+      ])))
+  def testBlockDiag(self, args):
+    args_maker = lambda: args
+    self._CheckAgainstNumpy(osp.linalg.block_diag, jsp.linalg.block_diag,
+                            args_maker, check_dtypes=True)
+    self._CompileAndCheck(jsp.linalg.block_diag, args_maker, check_dtypes=True)
+
 
   @parameterized.named_parameters(jtu.cases_from_list(
       {"testcase_name":
@@ -674,6 +845,7 @@ class ScipyLinalgTest(jtu.JaxTestCase):
       for dtype in float_types + complex_types
       for rng_factory in [jtu.rand_default]))
   @jtu.skip_on_devices("tpu")  # TODO(phawkins): precision problems on TPU.
+  @jtu.skip_on_flag("jax_skip_slow_tests", True)
   def testLuGrad(self, shape, dtype, rng_factory):
     rng = rng_factory()
     _skip_if_unsupported_type(dtype)
@@ -712,6 +884,8 @@ class ScipyLinalgTest(jtu.JaxTestCase):
   def testLuFactor(self, n, dtype, rng_factory):
     rng = rng_factory()
     _skip_if_unsupported_type(dtype)
+    if n == 200 and dtype == onp.complex64:
+      _skip_on_mac_xla_bug()
     args_maker = lambda: [rng((n, n), dtype)]
 
     x, = args_maker()
@@ -735,7 +909,7 @@ class ScipyLinalgTest(jtu.JaxTestCase):
       for lhs_shape, rhs_shape in [
           ((1, 1), (1, 1)),
           ((4, 4), (4,)),
-          ((8, 8), (8, 4, 2)),
+          ((8, 8), (8, 4)),
       ]
       for trans in [0, 1, 2]
       for dtype in float_types + complex_types
@@ -863,8 +1037,10 @@ class ScipyLinalgTest(jtu.JaxTestCase):
       for conjugate_a in (
           [False] if np.issubdtype(dtype, np.floating) else [False, True])
       for left_side, a_shape, b_shape in [
+          (False, (4, 4), (4,)),
           (False, (4, 4), (1, 4,)),
           (False, (3, 3), (4, 3)),
+          (True, (4, 4), (4,)),
           (True, (4, 4), (4, 1)),
           (True, (4, 4), (4, 3)),
           (True, (2, 8, 8), (2, 8, 10)),
@@ -886,6 +1062,33 @@ class ScipyLinalgTest(jtu.JaxTestCase):
                 unit_diagonal=unit_diagonal, left_side=left_side)
     jtu.check_grads(f, (A, B), 2, rtol=4e-2, eps=1e-3)
 
+  @parameterized.named_parameters(jtu.cases_from_list(
+      {"testcase_name":
+       "_A={}_B={}_bdim={}_leftside={}".format(
+           a_shape, b_shape, bdims, left_side),
+       "left_side": left_side, "a_shape": a_shape, "b_shape": b_shape,
+       "bdims": bdims}
+      for left_side, a_shape, b_shape, bdims in [
+          (False, (4, 4), (2, 3, 4,), (None, 0)),
+          (False, (2, 4, 4), (2, 2, 3, 4,), (None, 0)),
+          (False, (2, 4, 4), (3, 4,), (0, None)),
+          (False, (2, 4, 4), (2, 3, 4,), (0, 0)),
+          (True, (2, 4, 4), (2, 4, 3), (0, 0)),
+          (True, (2, 4, 4), (2, 2, 4, 3), (None, 0)),
+      ]))
+  def testTriangularSolveBatching(self, left_side, a_shape, b_shape, bdims):
+    rng = jtu.rand_default()
+    A = np.tril(rng(a_shape, onp.float32)
+                + 5 * onp.eye(a_shape[-1], dtype=onp.float32))
+    B = rng(b_shape, onp.float32)
+    solve = partial(lax_linalg.triangular_solve, lower=True,
+                    transpose_a=False, conjugate_a=False,
+                    unit_diagonal=False, left_side=left_side)
+    X = vmap(solve, bdims)(A, B)
+    matmul = partial(np.matmul, precision=lax.Precision.HIGHEST)
+    Y = matmul(A, X) if left_side else matmul(X, A)
+    onp.testing.assert_allclose(Y - B, 0, atol=1e-5)
+
   def testTriangularSolveGradPrecision(self):
     rng = jtu.rand_default()
     a = np.tril(rng((3, 3), onp.float32))
@@ -906,6 +1109,8 @@ class ScipyLinalgTest(jtu.JaxTestCase):
   def testExpm(self, n, dtype, rng_factory):
     rng = rng_factory()
     _skip_if_unsupported_type(dtype)
+    if n == 50 and dtype in [onp.complex64, onp.float32]:
+      _skip_on_mac_xla_bug()
     args_maker = lambda: [rng((n, n), dtype)]
 
     osp_fun = lambda a: osp.linalg.expm(a)
@@ -919,6 +1124,52 @@ class ScipyLinalgTest(jtu.JaxTestCase):
     self._CheckAgainstNumpy(osp_fun, jsp_fun_triu, args_maker_triu,
                             check_dtypes=True)
     self._CompileAndCheck(jsp_fun_triu, args_maker_triu, check_dtypes=True)
+
+  @parameterized.named_parameters(jtu.cases_from_list(
+    {"testcase_name":
+     "_n={}".format(jtu.format_shape_dtype_string((n,n), dtype)),
+     "n": n, "dtype": dtype}
+    for n in [1, 4, 5, 20, 50, 100]
+    for dtype in float_types + complex_types
+  ))
+  def testIssue2131(self, n, dtype):
+    _skip_on_mac_xla_bug()
+    args_maker_zeros = lambda: [onp.zeros((n, n), dtype)]
+    osp_fun = lambda a: osp.linalg.expm(a)
+    jsp_fun = lambda a: jsp.linalg.expm(a)
+    self._CheckAgainstNumpy(osp_fun, jsp_fun, args_maker_zeros,
+                            check_dtypes=True)
+    self._CompileAndCheck(jsp_fun, args_maker_zeros, check_dtypes=True)
+
+  @parameterized.named_parameters(jtu.cases_from_list(
+      {"testcase_name": "_lhs={}_rhs={}_lower={}".format(
+          jtu.format_shape_dtype_string(lhs_shape, dtype),
+          jtu.format_shape_dtype_string(rhs_shape, dtype),
+          lower),
+       "lhs_shape": lhs_shape, "rhs_shape": rhs_shape, "dtype": dtype,
+       "rng_factory": rng_factory, "lower": lower}
+      for lhs_shape, rhs_shape in [
+          [(1, 1), (1,)],
+          [(4, 4), (4,)],
+          [(4, 4), (4, 4)],
+      ]
+      for dtype in float_types
+      for lower in [True, False]
+      for rng_factory in [jtu.rand_default]))
+  def testChoSolve(self, lhs_shape, rhs_shape, dtype, lower, rng_factory):
+    rng = rng_factory()
+    _skip_if_unsupported_type(dtype)
+    def args_maker():
+      b = rng(rhs_shape, dtype)
+      if lower:
+        L = onp.tril(rng(lhs_shape, dtype))
+        return [(L, lower), b]
+      else:
+        U = onp.triu(rng(lhs_shape, dtype))
+        return [(U, lower), b]
+    self._CheckAgainstNumpy(osp.linalg.cho_solve, jsp.linalg.cho_solve,
+                            args_maker, check_dtypes=True, tol=1e-3)
+
 
 if __name__ == "__main__":
   absltest.main()
